mtd: fixing CFI for Numonyx CFI bug
commit  b8181c701f81bee9f99d14bdf8d4b800e49460be from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The M29EW devices seem to report the CFI information wrong when
it's in 8 bit mode.

There's an app note from Numonyx on this issue and there's a patch
in the open source, but it doesn't seem to be in the mainline.

This commit is based on that patch from Massimo Cirillo titled:
buffer size adjustment for M29EW Numonyx devices.

Signed-off-by: John Linn <john.linn@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>