FogBugz #177284: add Altera VIP framebuffer driver
commit  dae620d9475f829c9fd374f27850998c3ddbb629 from
https://github.com/altera-opensource/linux-socfpga.git

This adds support for the Altera framebuffer that uses the Altera VIP
Frame Reader hardware.

The current implementation of the Altera VIP driver was sitting in the
wrong directory. It was sitting in drivers/video, and should reside inside
of /drivers/video/fbdev.

Signed-off-by: Chris Rauer <crauer@altera.com>
Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>