rpmsg: rpc: use the local device pointer in all file operations
commit 83f7a0b8963fd4181c3e63605deef822e8c87ed9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The remote processor recovery process includes the deletion and
recreation of an rpmsg-rpc device. The representative rppc_device
structure is retained and reused if there are any open applications
using the exposed character device. The underlying device pointer
for a rppc_device is though deleted and recreated and can become
NULL at any point if an error recovery happens. So, switch to using
the local reference device pointer in all the fop functions for
the exposed character device.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>