MLK-10170 usb: phy: mxs: keep USBPHY2's clk always on
commit 28e487f5c6d2dc3596b31f270b10d61f65839149 from
https://source.codeaurora.org/external/imx/linux-imx

Per IC engineer request, we need to keep USBPHY2's clk always on,
in this way, the USBPHY2 (PLL7) power can be controlled by
hardware suspend signal totally. It is benefit of USB remote wakeup
case which needs the resume signal be sent out as soon as
possible (without software interfere).

It is intended to fix the issue which this ticket describes, the
reason for this issue is the host does not send resume in time.

Signed-off-by: Peter Chen <peter.chen@freescale.com>
(cherry picked from commit 98888b352377f9ebaee03bedce8c239691f45262)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>