MLK-10959: mxc: sim: Disable cwt timer when tx and enable in rx.
commit 59ab52a2497d7296e9cc95b4d0e560c470ecd3e5 from
https://source.codeaurora.org/external/imx/linux-imx

The CWT timer is used to detect the the character interval in the data traffic.
When tx, SIM IP can guarantee the interval based our setting. When RX, we need
to enalbe the CWT timer to check whether the interval is in the range. This patch
fix this.

Signed-off-by: Luwei Zhou <b45643@freescale.com>
(cherry picked from 9c92dfd070e7427eb1e0166f368b89b4a7ac1bff)