MLK-12185 mlb: imx: fix the operation for mlb status in mlb_isr
commit 10b8c40353676c6c0be30fec672589c805c1ab11 from
https://source.codeaurora.org/external/imx/linux-imx

As a result that the wrong operation is used for the mlb status
in mlb_isr(), some results are independent of their operations.

for example:
rx_cis = (cdt_val[2] & ~MASK) >> SHIFT, where, MASK = 0xf0000000
and SHIFT = 28. So, the result is always 0 regardless of the values
of its operands.

This patch fixes the operation for mlb status in mlb_isr().

(reported by coverity check)

Signed-off-by: Gao Pan <pandy.gao@nxp.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>