MLK-12372: ASoC: fsl_asrc: fix dma task timeout after suspend/resume
commit 2c839db3aaef75cc0ffaaae525e89c979a8d03e3 from
https://source.codeaurora.org/external/imx/linux-imx

commit 743cead0f8c4ac6311ffb500efd6146c40124310 is not a complete fix.
There is low possibility that this issue still occur.

Last commit add init_completion() in the suspend function, but if the
dma callback function is called after convert error, the complete is
done, the init_completion will not be called, so the complete state is
not correct in the next conversion.

This patch is to move init_completion to the beginning of conversion
to fix the issue.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>