dts: s32v: DRAM correct size of DDR regions in dts and dtsi files
commit f2c2e9c16307c13a2c62a1f53d25b7df4728784d from
https://source.codeaurora.org/external/autobsps32/linux

The size of DDR regions was corrected in dts files (dtsi file of evb,
because this is used by more evb dts files). In the SoC dtsi file,
it is used the size of DDR than can be mapped.
The board dts files will then overwrite these settings with their
available DDR size.

Issue: ALB-5945

Signed-off-by: Ana Nedelcu <ana.nedelcu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>