ALB-241 Update the dtsi in order to work with both DDRs
commit  1bd99357359ccabfaa0e06fa5f2d2031595f90f1 from
https://source.codeaurora.org/external/autobsps32/linux

The current version of u-boot is using DDR0 instead of DDR1,
so the dtsi has to be updated in order to work with this version
of u-boot.

The Spintable addresses and the memory reserved zone are
specified in DDR0 instead of DDR1.

Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>