clocksource: fsl_stm: Add STM driver support
commit 9717942c4070f79f6349fbf143cbbc0189653f9d from
https://source.codeaurora.org/external/autobsps32/linux

The driver can be used with any two cores.
STM is not suitable for using it as a clocksource because the
counter is shared between channels so it is shared with the
channel used for clockevents. To avoid losing events, the counter is
stopped in the event interrupt, leading to inaccurate values for the
clocksource.

Issue: ALB-420

s32gen1: stm: Fix STM not appearing in menuconfig.

The STM driver was not appearing in menuconfig because it laked a proper
description in the kconfig file. Modified the Kconfig file so that a
string is provided to be used by menuconfig in order to be have the
configuration option shown.

Issue: ALB-5093

clocksource: fsl_stm: Rename the file for consistency

For the sake of consistency, let's rename the file to a name similar
to other file names in this directory.

Issue: ALB-5454

clocksource: fsl_stm: Optimize STM resources allocation

STM module allocates stm structures for each possible core. This can be
optimized in order to allocate only as many structures as stm modules
defined in dts.

Issue: ALB-471

clocksource: fsl_stm: Upgrade to 4.14 & 5.10

Fix STM driver build errors by migrating to the new 'set-state'
interface and hotplug state machine.

Issue: ALB-2714
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Ghennadi Procopciuc <Ghennadi.Procopciuc@nxp.com>
Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>