STM driver support
commit  d4a11d246a2fb808ed83e78fcdd33ebbc0d138d4 from
https://source.codeaurora.org/external/autobsps32/linux

The driver can be used any two cores.
STM is not suitable for using it as a clocksource because the
counter is shared between channels so it is shared with the
channel used for clockevents. To avoid losing events, the counter is
stopped in the event interrupt, leading to inaccurate values for the
clocksource.

Issue: ALB-420

Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>