tty: serial: linflexuart: Add baud rate programming
commit 1a11c59d4d3ce39fbbeb58090dc7bd351ca11f5c from
https://source.codeaurora.org/external/imx/linux-imx

Baud rate is computed as LIN_CLK / (OSR × LDIV) when ROSE is 1 or
LIN_CLK / (16 × LDIV) when ROSE is 0. LDIV is an unsigned fixed point
number. The mantissa is coded into 20 bits of LINIBRR and the fraction is
coded on 4 bits of LINFBRR, so setting the baud rate means setting LINFBRR
and LINIBRR.

LINIBRR is the integer part of LIN_CLK / (f(ROSE) * baud rate) and LINFBRR
is the fractional part of it multiplied by 16, where f(ROSE) can be 16 or
OSR value.

Signed-off-by: Adrian.Nitu <adrian.nitu@freescale.com>
Signed-off-by: Stoica Cosmin-Stefan <cosmin.stoica@nxp.com>
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Reviewed-by: Fugang Duan <fugang.duan@nxp.com>