Input: goodix - put reset and irq into default state during probe
commit 2305d0c92c8b98f4828b7d35a6e46905fef5a7e2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The Goodix touchscreen controller requires a specific pin sequencing to be
done when the device is out of reset. Don't assume the initial state of the
pins at probe. Explicitly put the touchscreen controller into a reset state
and set the interrupt pin to low.

Also don't allow reset pin to float by configuring it as an input. This
avoids assumptions based on board's hardware.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>