s32v234_pcie: Run-time detection of chip revision to ignore ERR009852
commit  b77ced4649450dd7fb8259fd8317c153663a668d from
https://source.codeaurora.org/external/autobsps32/linux

If the chip revision is new enough, PCIe erratum ERR009852 can be safely
ignored regardless of the corresponding Kconfig choice. This is in fact
done by u-boot, which inserts a property in the fdt for us to parse.
If the revision cannot be inferred, fall back to the Kconfig choice.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-876
Signed-off-by: Meng Li <Meng.Li@windriver.com>