ASoC: tlv320aic3x: Add controls for selecting HP power down modes
commit c0aa990197a2e754dfbee1dec73122f7520dec1c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When the HP drivers are powered down, their output can be placed in
high-impedance state or it can be weakly driven to the VCM level.
High-impedance mode provides better power saving in the expense of pop
noise when powering up the HP.
When the output is driven to VCM level the artifacts can be minimized, but
the power down power consumption will be slightly higher.

The patch adds way to select the HPL/R power-down mode.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>