MLK-13638-4 regulator: fixed: add system pm routines for pinctrl
commit f8ed5516d54292d0823d9d99b141e1b7dbd38005 from
https://source.codeaurora.org/external/imx/linux-imx

At some systems, the pinctrl setting will be lost or needs to
set as "sleep" state to save power consumption. So, we need to
configure pinctrl as "sleep" state when system enters suspend,
and as "default" state after system resumes. In this way, the
pinctrl value can be recovered as "default" state after resuming.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>