MLK-14258: ASoC: imx-pcm-rpmsg: sync sample rate with tx and rx
commit 5d8c7ea7ce3e6f36d26e5125f5e8e13948c2fb83 from
https://source.codeaurora.org/external/imx/linux-imx

In suspend and resume, the M4 side will reset hw parameter for tx
and rx, when only tx is working, the parameter of rx is a old value,
which will cause the parameter is not sync with tx and rx.
currently the M4 audio can only work in sync mode, so set both
parameter in same time.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>