MLK-14396 usb: chipidea: add mutex for concurrent running
commit cebd6f46dd0545716b8d730de284be6349eb7a00 from
https://source.codeaurora.org/external/imx/linux-imx

The ci_handle_id_switch is called at two places, at very rare situations,
it may be running at the same time. Eg, when the system is back from
the resume, the id event is occurred from extcon driver, as well as
power_lost work item is called due to the controller is poweroff at
the suspend.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>