MLK-14582: ASoC: imx-pcm-rpmsg: fix audio noise issue with pulseaudio
commit 14e0dcede5381bcde8308f9b28cf0896107b3c6f from
https://source.codeaurora.org/external/imx/linux-imx

The pulse audio will set a wrong buffer size which is not the integral
multiple of period size, it will cause the DMA can't work correctly in
M4 side.
The reason is that we always need to add a constraint for
SNDRV_PCM_HW_PARAM_PERIODS, which make it integer.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>
(cherry picked from commit 54a10a6c2130a69aca4c1923dac3a15137911146)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>