MLK-15296-3: ASoc: fsl: remap the dsp firmware
commit 889abf70e5d5fad4613c940a8cf3335ddebaf628 from
https://source.codeaurora.org/external/imx/linux-imx

In order to use the hifi4's Cache to cache the firmware's
.rodata, .text, .data, .bss section and hifi4 core lib's
.text section, the firmware's .rodata, .text, .data and
.bss section should be remaped to 0x20700000 - 0x20FFFFFF
address range. This patch is used to parse the firmware
and load each section to corresponding address range.

This patch also set csr_gpr_control to 0x515A2080 to
remap the hifi4's address range in SCFW.

In addtion, add cases to support hifi4 framework's
performance test.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>