MLK-15927-1: ASoC: fsl_sai: Fix noise when using EDMA
commit 90efe83b6e9cc0b77fd9f117035fdfcf46b29dc2 from
https://source.codeaurora.org/external/imx/linux-imx

EDMA requires the period size to be multiple of maxburst. Otherwise the
remaining bytes are not transferred and thus noise is produced.

We can handle this issue by adding a constraint on
SNDRV_PCM_HW_PARAM_PERIOD_SIZE to be multiple of tx/rx maxburst value.

This is based on a similar patch we have for ESAI:
commit bd3f3eb2a37c
("MLK-15109-2: ASoC: fsl_esai: add constrain_period_size")

Signed-off-by: Mihai Serban <mihai.serban@nxp.com>
Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>