MLK-16007-2: ASoC: fsl_asrc: Enable automatic configuration for P2P
commit 77924f38da9d811ca20d393294e9d5be6a0150ed from
https://source.codeaurora.org/external/imx/linux-imx

Use automatic selection of processing options and internal measured
ratio for P2P conversions.

The conversion done by ASRC depends on the IPG master clock frequency
that can have any value between 130MHz and 200MHz. The documentation
states that to support 10 channels with 192KHz sampling rate the
master clock frequency must be at least 160MHz.
When the master clock cannot be configured to faster frequencies the
ASRC can still convert the samples but it have to be configured to
automatically adjust the processing options and conversion ratio.

Signed-off-by: Mihai Serban <mihai.serban@nxp.com>
Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>