MLK-16203 enable runtime pm of i2c temporary when do system suspend
commit 3972bc5c1214ec029a53890e5b4c3a756d3eca03 from
https://source.codeaurora.org/external/imx/linux-imx

When we do system suspend, the runtime pm will be disabled, but we need
to control the PMIC to power on/off the regulator, if the runtime pm is
disabled, if will failed to request runtime wakeup. So data transfer will
failed.

Signed-off-by: Bai Ping <ping.bai@nxp.com>
Signed-off-by: Gao Pan <pandy.gao@nxp.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>
(cherry picked from commit 93adab71408f1f162015e77715dd04ce2301d673)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>