i2c: imx: implement bus recovery with gpio for Layerscape
commit 90b29e7b9c6bd6f9c2e27a2275d7c68d80cedf42 from
https://source.codeaurora.org/external/imx/linux-imx

Based on the I2C specification, if the data line (SDA) is stuck low,
the master should send nine clock pulses. The I2C slave device that
held the bus low should release it sometime within those nine clocks.

Because pinctrl is not supported on Layerscape, current bus recovery
is not avalible for Layerscape. This patch uses an open drain GPIO
pin to connect to the IICx_SCL to drive nine clock pulses to unlock
the I2C bus.

Signed-off-by: Zhang Ying-22455 <ying.zhang22455@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>