MLK-17152: ASoC: fsl_hifi: support suspend and resume
commit 78d2651d92f99e5c9e1520523315740a915e3b7b from
https://source.codeaurora.org/external/imx/linux-imx

For hifi need to enter runtime suspend state in suspend,
then the power of HIFI can be down. In this case content
in internal RAM will be lost, and need to be recovered
in resume.

Move the loading firmware to runtime resume function, and
define ICM_SUSPEND and ICM_RESUME command, with ICM_SUSPEND
the hifi framework will store the data in RAM and with
ICM_RESUME the hifi framework will restore the data to RAM.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>