MLK-17309-4: ASoC: fsl_hifi: fix crash issue when destination not align
commit 1ee21d58911dcaa125e063b8c82db7328549f22c from
https://source.codeaurora.org/external/imx/linux-imx

When loading the codec libs in driver, if the destination is
not 4-bytes alignment when doing memset_hifi(), the driver
will print a warning message and the driver may crash in some
cases.

So by changing the memset() function and aligning the virtual address
based on the physical address to fix this issue.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>