MLK-17309-4: ASoC: fsl_hifi: fix crash issue when destination not align
commit 7d877756ebf7732bba0b08e97e93d10a297587d0 from
https://source.codeaurora.org/external/imx/linux-imx

When loading the codec libs in driver, if the destination is
not 4-bytes alignment when doing memset_hifi(), the driver
will print a warning message and the driver may crash in some
cases.

So by changing the memset() function and aligning the virtual address
based on the physical address to fix this issue.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>