MLK-17312-5 usb: cdns3: improve the role switch process
commit 120b51228b7aaace5582d857fb92cef4ef497822 from
https://source.codeaurora.org/external/imx/linux-imx

Current design tries to switch role no matter it is a dual-role device
or a single-role device. It produces extra switch process, and have an
error message at console when tries to start a non-exist role.

In this commit, we do below changes
- The role switch work queue is only for dual-role or peripheral-only
device.
- For peripheral-only device, we need to switch role to CDNS3_ROLE_END
since we need to close vbus and turn off clocks at this role when the
cable is disconnected from the host; And we do noop when the external
cable indicates we are host.

Acked-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
(cherry picked from commit 2000169e689ef955ec312f84976c098519487890)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>