TEMP: dt-bindings: soc: ti: Add Keystone DSP Memory mapping binding
commit dda3976a6d4cc0d9f48798fce6e901f9a3399f30 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the device tree binding document for the userspace memory mapping
driver on TI Keystone SoCs. This driver is used to expose either some
portions of the DDR and/or some portions of the On-chip RAM controlled
by the Multicore Shared Memory Controller (MSMC) to userspace for the
Multi Proc Manager (MPM) stack.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>