drm: xlnx: drv: Allow multiple masters on a drm device
commit  2b57fd0861296de06fa0333a8b62c153e0991726 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This may not be the ideal solution, but is to allow multiple
masters running on a single drm device, when the application
has the admin priviliege. This enables access to master
only resources such as setting drm planes and etc.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>