drm/omap: Add ability to check if requested plane modes can be supported
commit dcad650de1870bb465aba7b383fb8f11f346dfe2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

We currently assume that an overlay has the same maximum width and
maximum height as the overlay manager. This assumption is incorrect. On
some variants the overlay manager maximum width is twice the maximum
width that the overlay can handle. We need to add the appropriate data
per variant as well as export a helper function to retrieve the data so
check can be made dynamically in omap_plane_atomic_check().

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>