dpaa_eth: fix iova handling for sg frames
commit 44fffdeb55f924392763a4d0f7e7a887e5876302 from
https://source.codeaurora.org/external/imx/linux-imx

The driver relies on the no longer valid assumption that dma addresses
(iovas) are identical to physical addressees and uses phys_to_virt() to
make iova -> vaddr conversions. Fix this also for scatter-gather frames
using the iova -> phys conversion function added in the previous patch.
While at it, clean-up a redundant dpaa_bpid2pool() and pass the bp
as parameter.

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>
Acked-by: Madalin Bucur <madalin.bucur@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>