drivers/soc/fsl: add EPU FSM configuration for deep sleep
commit 08dcfadc9a22bdb8cd37bc1ce38bc98d202cd857 from
https://source.codeaurora.org/external/imx/linux-imx

In the last stage of deep sleep, software will trigger a Finite
State Machine (FSM) to control the hardware procedure, such a
board isolation, killing PLLs, removing power, and so on.

When the system is waked up by an interrupt, the FSM controls
the hardware to complete the early resume procedure.

This patch configure the EPU FSM preparing for deep sleep.

Signed-off-by: Hongbo Zhang <hongbo.zhang@freescale.com>
Signed-off-by: Chenhui Zhao <chenhui.zhao@freescale.com>
Signed-off-by: Ran Wang <ran.wang_1@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>