staging: fsl_ppfe/eth: support for userspace networking
commit 5c20cca33475358a3416ab6084ca5d1a6e7c7e23 from
https://source.codeaurora.org/external/imx/linux-imx

This patch adds the userspace mode support to fsl_ppfe network driver.
In the new mode, basic hardware initialization is performed in kernel, while
the datapath and HIF handling is the responsibility of the userspace.

The new command line parameter is added to initialize the ppfe module
in userspace mode. By default the module remains in kernelspace networking
mode.
To enable userspace mode, use "insmod pfe.ko us=1"

Signed-off-by: Akhil Goyal <akhil.goyal@nxp.com>
Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>