MLK-18279: ASoC: fsl_dsp: get the information of reserved memory from dts
commit dc9756fa77434d5a3c85ce44a82565c2fe6b89ab from
https://source.codeaurora.org/external/imx/linux-imx

The reserved memory for dsp is defined in dts file, however, the dsp
driver has also defined the address and size of this reserved memory,
which is repeated and inflexible.

So by cancelling the definition in dsp driver and use system API to
get the information of reserved memory from dts dynamically to fix
this problem.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>