MLK-18279: ASoC: fsl_dsp: get the information of reserved memory from dts
commit f83bf722943d5f54689b0ff5635f5605cf73f598 from
https://source.codeaurora.org/external/imx/linux-imx

The reserved memory for dsp is defined in dts file, however, the dsp
driver has also defined the address and size of this reserved memory,
which is repeated and inflexible.

So by cancelling the definition in dsp driver and use system API to
get the information of reserved memory from dts dynamically to fix
this problem.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>