MLK-18319-2 usb: cdns3: gadget: delete the unless code
commit b8407f59c070232be5c31c242b0343837a4ee685 from
https://source.codeaurora.org/external/imx/linux-imx

Confirmed from CDNS, this code is from bare metal, and not needed
for Linux. Below are their response:

This is alternative deferred interrupt.
This is intended to be used in Bare Metal to not block system in
interrupt.
This was not fully checked in Linux, so we do not recommend to use it
there.

Reviewed-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
(cherry picked from commit e5a1abc937692efe1cc7347ae69a31e80b751352)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>