drm: xlnx: sdi: Fix the VSync_Start and VSync_End for interlaced modes
commit  734a066c195884bd7f3de8ba6f516994987aa50e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Fixes the VSync_start and VSync_End for Vertical Sync Line for Field 0
and Field 1 for interlaced modes. The Field 0 is to be incremented by 1
for all interlaced modes. The Field 1 is to be incremented by 2 for 3GB
modes and by 1 for all other interlaced modes.
This can't be fixed in any other way by modifying the timing values in
the modes array.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>