linflex: Don't disable tx and rx during shutdown
commit  d9a6daaf67d63362e6f8140b2f1bde94203481d8 from
https://source.codeaurora.org/external/autobsps32/linux

The issue is reproducible on simulator which is slower than the
hardware.
The driver waits for ack from edma which signals when all characters
were put in linflex fifo. That isn't equivalent
with that fact that linflex fifo is empty at that moment. If
shutdown is called, before fifo is empty, the characters from the fifo
will not be sent anymore.

Issue: ALB-2766
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>