MLK-19151: ASoC: fsl_asrc: fix dma timeout issue for imx8qxp
commit 500a34e3246e15fdd5a1408d271d32112ae37930 from
https://source.codeaurora.org/external/imx/linux-imx

In the imx8qxp, the DMA is EDMA, which require the buffer size
should be divided by burst size with no remainder, otherwise
the remainder is not transferred by EDMA, so the input buffer
is not consumed by ASRC, then there will be dma output timeout
issue. This behavior is different with SDMA.

This patch is to change the input burst size to be 1 to avoid
the issue, and refine the last_period_size for output buffer.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>