MLK-19152-2 drm/imx: lcdif: force 'mode_changed' when fb width changed
commit f5ca7971de27682dc54ea14a06b5d8a15fe81c89 from
https://source.codeaurora.org/external/imx/linux-imx

In DRM atomic modeset check, it will not check the fb's width
change, so in later atomic commit, it will not disable the CRTC
which has no mode changed. But for LCDIF, the fb width related
registers configuration can not be done when LCDIF is running.
So force 'mode_changed' to be true when fb width changed.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(cherry picked from commit 518ff82756a39ff2d2f750596295baa4f5fca4c5)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>