MLK-19152-1 gpu: imx: lcdif: realize fb horizontal crop via Pigeon Mode
commit e5c67c1cef03d6ebe565c3088b1bd3507f7290f8 from
https://source.codeaurora.org/external/imx/linux-imx

According to the LCDIF specification, the Legacy Mode does not
support cropping function in the horizontal direction, so add
Pigeon Mode which can support this kind of function. And when
enable this mode, the legacy horizontal timings configuration
should use stride value but not the active width, and related
pigeon configuration should use the active width but not the
stride value.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(cherry picked from commit e6da9542693dd585972897f62748a101f5726a74)