sdhci-esdhc-imx: Use MMC_CAP2_NO_WRITE_PROTECT if ESDHC_WP_NONE
commit 9498694f34da611068c0630bc3ab242f9adb963d from
https://source.codeaurora.org/external/autobsps32/linux

If the 'wp_type' field of 'struct esdhc_platform_data'
is left as ESDHC_WP_NONE, meaning there is no information in the
device tree about accessing and reading the state of the write-protect
switch, then MMC_CAP2_NO_WRITE_PROTECT should be added to the list
of capabilities.

As a result, the mmc core is aware of the situation and does not
report a "host does not support reading read-only switch,
assuming write-enable" warning when unable to read the status
of the write-protect switch.

Issue: ALB-2885
Signed-off-by: Dan Nica <dan.nica@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>