MLK-18979-1: ASoC: fsl_asrc: add resume function for asrc_m2m
commit e620dc8f1bec329f38c766efb4bbd83f6b7176b7 from
https://source.codeaurora.org/external/imx/linux-imx

There will be "output DMA task timeout" after suspend and resume.
The reason is there is not enough data in the input FIFO.

In the fsl_asrc_start_pair function we initialize the FIFO with
zero data after pair is enabled, it looks like we add more data
to input FIFO. For example if the input buffer length is 100,
but the actual length is 100 + channel*4. so we need to do same
work in resume for the asrc pair is disabled in suspend, the
input FIFO is cleared.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>