MLK-19565-1: ASoC: fsl_rpmsg_i2s: support rpmsg on imx8qm
commit 26e4c3d80f3ce07dbe5732009b4e6a46ec370966 from
https://source.codeaurora.org/external/imx/linux-imx

On imx8qm mek, the cs42888 is connected with i2c in cm41 domain,
but wm8960 is connected with i2c1, which is not in m4 domain.
So we only need to eable rpmsg for cs42888.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
(cherry picked from commit 9d2368aef40e4d107e4deee1a2c7e191c1afe644)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>