net: networkproxy: add network proxy framework
commit b126aaa6155f7fe9604cda1132bfac4572632862 from
https://github.com/intel/linux-intel-lts.git

Network proxy maintains network presence while Host is in low-power mode.
Host is woke-up when there is print job or other packet of interest not
handled by Network Proxy agent.

This technology is described in ECMA-393 proxZZzy. It requires 2 devices
to be registered to the Network Proxy framework: IPC Device and Network
Device.

Signed-off-by: Ong, Boon Leong <boon.leong.ong@intel.com>
Signed-off-by: Lay, Kuan Loon <kuan.loon.lay@intel.com>
Signed-off-by: Song, Yoong Siang <yoong.siang.song@intel.com>
Signed-off-by: Weifeng Voon <weifeng.voon@intel.com>
Signed-off-by: Voon Weifeng <weifeng.voon@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>