LF-811-1: drm/bridge: Add driver for legacy Freescale Seiko 43WVFIG adapter
commit c40c90f2d3265167edd6384dd8691dc88f78308d from
https://source.codeaurora.org/external/imx/linux-imx

This is an adapter card made for the 4.3", 800x480, LCD panel Seiko
43WVFIG. The LCD panel is a 24bit DPI bus, while the adapter card has
two ports: 18-bit and 24-bit data input. For the 18-bit data input, the
adapter card is demuxing some of the data lines, in order to feed all of
the 24 lines needed by the LCD.
This driver handles both this use-cases.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>