MLK-19819-1 drm/imx: lcdif: bypass atomic check when CRTC will be disabled
commit b9aefecc286b3564dc40e1cef962712aa4c33850 from
https://source.codeaurora.org/external/imx/linux-imx

On 4.14.y kernel branch, the DRM framework has been modified that
when no CONNECTOR attach to CRTC, the fb creation wil be deferred
until some CONNECTOR has been detected via hotplug. And the system
suspend workflow is also affected accordingly, if the CRTC atomic
check fails, the display-subsystem suspend also will be caused to
fail. So bypass the 'bus_format' check when CRTC is going to be
disabled.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>