dmaengine: xilinx_frmbuf: Mark previous active decriptor as complete in irq handler
commit  16851548e03584facc0e4c523a00efb62204be66 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Moving the block to mark previous active descriptor as complete
to irq handler as it's more appropriate to do it in interrupt
context which confirms that previous transaction was completed.

There is no semantic change in the flow,
previously descriptor was marked as completed in
xilinx_frmbuf_start_transfer() assuming it's is called by
the done interrupt handler which didn't seem required,
so we now mark it as completed in the interrupt handler itelf.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>