MLK-20277-3: ASoC: imx-pcm-rpmsg: change the state of substream in resume
commit 93e25da90ff2bd7db370dd6194ab84e78ad2ef8a from
https://source.codeaurora.org/external/imx/linux-imx

In LPA mode, the system will be resumed by audio notification, when the
period size is small, there will be occasion that when notification
the underrun is happen, but the substream runtime state is not running
so the aplay won't trigger stop first, then start. just only trigger
the start, which don't comply with the convention.
So in this patch, we change the substream runtime state to running, when
the notification happened at resume.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>