MLK-20428-2: ASoC: imx-cs42xx8: The FMT should be same for tx and rx
commit 4dcdcf9d6b562ae8a1f7c66ab337dd6c0a554a1d from
https://source.codeaurora.org/external/imx/linux-imx

If the FMT of tx and rx is different, there will be issue. for example
tx is working at TDM mode, but rx want to work at I2S mode, this case
is not supported in current ALSA driver, the best way to support this
is to distinguish substream in set_fmt function.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>