drm/mxsfb: Improve the axi clock usage
commit 34ac946b631c67b53ef70aaf8b3856fe5d3a3b43 from
https://source.codeaurora.org/external/imx/linux-imx

Currently, the enable of the axi clock return status is ignored, causing
issues when the enable fails then we try to disable it. Therefore, it is
better to check the return status and disable it only when enable
succeeded.
Also, remove the helper functions around clk_axi, since we can directly
use the clk API function for enable/disable the clock. Those functions
are already checking for NULL clk and returning 0 if that's the case.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Acked-by: Leonard Crestez <leonard.crestez@nxp.com>
Tested-by: Guido Günther <agx@sigxcpu.org>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>