GPIO: Add GPIO driver to S32-gen1
commit  b4cab00489777f0749b7f348d9ce793e6099d2e8 from
https://source.codeaurora.org/external/autobsps32/linux

Add GPIO driver for S32-Gen1 to allow multiple instances of the driver
to share the same register ranges due to the non-contiguous
arrangement of the EIRQ pins. The device tree binding of
this driver will be changed to use the regmap API to allow
this sharing of resources.

Issue: ALB-3170

Signed-off-by: Matthew Nunez <matthew.nunez@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>