arm64: dts: ti: k3-am65-mcu: Reserve some MCU SRAM for MCU R5F0
commit 69ed7e5824deae0e3a68586277f015af19a45dbd from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a child SRAM node to reserve a portion of the MCU domain on-chip
SRAM memory for use by the MCU domain R5F Core0. A preliminary size
of 256 KB (half of the current SRAM) is reserved to begin with and
can be adjusted as per need.

This memory is also added to the R5F0 remoteproc node using the
sram property.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>