drm/omap: add crtc transparency key property
commit 1bc5306b0bca8630bbe01b8d3323a5e593c841d6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add support for transparency keying.

User space must understand how the TCK works and needs to place the planes
to correct zpos. The generic zpos normalization in DRM however is going
to do normalization and might move the position of the planes which breaks
the TCK as planes are no longer in the position where application
deliberately placed them.

If the TCK is enabled for the crtc then use the zpos configured by the
application instead of the normalized positions and at the same time do a
sanity check against overlapping zpos.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>