usb: hub: fix over-current race condition
commit  7ddc88e5fa32879d1fb2d7deae357899b54f8b24 from
https://github.com/axxia/linux-yocto.git v5.10/standard/preempt-rt/axxia-dev/base

There exists a race condition when an over-current fault is cleared,
between the event processing in the hub_event function and the setting
of the event bits in the hub_irq function. When the over-current is
cleared the hub_event call tries to read the event bits, which havent
been set by the hub_irq.  In this patch when the over-current is
cleared the hub_event will check the event bits and if they are not
set it will pause and check again. It will sync the event bits.

Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Charlie Paul <charlie.paul@windriver.com>
Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>