mmc: sdhci-esdhc-imx: do not enable wakeup by default
commit 50f70ed62e6e13243049aa2f4796680159ca691e from
https://source.codeaurora.org/external/imx/linux-imx

After adding mega fast support, the default enabled usdhc wakeup will block
M/F to gate off power domain.

To avoid this issue, we only claim wakeup capability and reply on user to enable
it via sysfs according to real needs.
The drawback of such change is that for SDIO WiFi Wakeup On Wireless feature,
User has to enable both uSDHC and WiFi WoW wakeup mannually to make
WoW work well.

BTW, due to the wakeup feature is controller itself, so we do not need to reply
on WiFi PM flags to enable it.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>