arm64: dts: lx2160aqds: Add mdio mux nodes
commit 6d5d58cbd5f0c4c65a16e31fb5de3a2c4718abfd from
https://source.codeaurora.org/external/imx/linux-imx

The two external MDIO buses used to communicate with phy devices that are
external to SOC are muxed in LX2160AQDS board.
These buses can be routed to any one of the eight IO slots on LX2160AQDS
board depending on value in fpga register 0x54.
Additionally the external MDIO1 is used to communicate to the onboard
RGMII phy devices.
The mdio1 is controlled by bits 4-7 of fpga register and mdio2 is
controlled by bits 4-7 of fpga register.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>