uio: Add dma-buf import ioctls
commit  d24314e88312f655baa5437a45258811e80b01cb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Add the dmabuf map / unmap interfaces. This allows the user driver
to be able to import the external dmabuf and use it from user space.

As noted in the header, this is an experimental feature, so it may
change or deprecate at any time.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Link: https://lkml.org/lkml/2019/2/23/207
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>