net: genet: enable link energy detect powerdown for external PHYs
commit  1fd2646eb19c4b9d395b68c00599eca03fff83e7 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

There are several warts surrounding bcmgenet_mii_probe() as this
function is called from ndo_open, but it's calling registration-type
functions. The probe should be called at probe time and refactored
such that the PHY device data can be extracted to limit the scope
of this flag to Broadcom PHYs.

For now, pass this flag in as it puts our attached PHY into a low-power
state when disconnected.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>