drm/imx: sec-dsim_imx: correct DSIM version register access
commit 2d489fafab7979e9e374daefb4432d16a1082cb3 from
https://source.codeaurora.org/external/imx/linux-imx

During the DSIM binding stage, only enable cfg clock is not
enough to access the version register, since at this time,
if the DSIM is still in the reset state, the register read
will always return 0. So before the version register read,
the runtime pm should be in resume state.

(This is the DSIM controller driver part change.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>