MLK-21880-1 drm/bridge: sec-dsim: correct DSIM version register access
commit 7b01081fad884a5f30bff733c26bcd94de80dd99 from
https://source.codeaurora.org/external/imx/linux-imx

During the DSIM binding stage, only enable cfg clock is not
enough to access the version register, since at this time,
if the DSIM is still in the reset state, the register read
will always return 0. So before the version register read,
the runtime pm should be in resume state.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>