MLK-22210 mxc: emvsim: enable ICM mode
commit c79402071ae74826ce35cacf6d55b89fe614a7d9 from
https://source.codeaurora.org/external/imx/linux-imx

Once detected a valid TS, the ICM bit is cleared and the data format bit,
IC bit , is set to appropriate value depending on the data format
detected using the initial character.

When TS is 0x3B (direct convention) in cold reset, and then become 0x3F
(inverse convention) in warm reset, it will not in ICM mode.

Signed-off-by: Joakim Zhang <qiangqing.zhang@nxp.com>
Acked-by: Fugang Duan <fugang.duan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>