reset: Add driver for dispmix reset
commit 6d73c7f3ae06752a6f2afd0815f436a3e8079c05 from
https://source.codeaurora.org/external/imx/linux-imx

This is an reset driver to implement a reset controller
device DISPMIX on IMX8MM and IMX8MN platforms. Dispmix
reset is used to reset or enable related buses and clks
for the submodules in DISPMIX.

All the dispmix resets are divided into three subgroups:
sft_rstn, clk_en and mipi_rst, and each of them contains
several reset lines to control several different modules
on and off in DISPMIX which doesn't require the standard
reset flow, but only line assert and deassert operations.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>