MLK-18605-12 ARM64: dts: imx8mm-evk: add support for RM67191 panel
commit 3c3dac36a2c673148290f3d316419b6a1a13aacb from
https://source.codeaurora.org/external/imx/linux-imx

Create a new dts 'fsl-imx8mm-evk-rm67191.dts' to support panel
'RM67191' display which is attached to DSIM controller directly.
So the corresponding panel device node is defined as the child
of 'mipi_dsi' node under the DRM DSI framework. Since the
'adv_bridge' and 'RM67191' should be enabled exclusively, so
disable 'adv_bridge' when enable 'RM67191'.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>