MLK-22355: mfd: si476x: Use system_freezable_wq instead of system_wq
commit 1ae9705fc3a81989b570d1981bc3abb56e4ae144 from
https://source.codeaurora.org/external/imx/linux-imx

There is issue that system can't enter suspend while the si476x is
working.

The reason is that with the workqueue thread is still working after
i2c enter suspend, then cause the cpu_suspend function failed.

This patch is to use the system_freezable_wq instead of the system_wq,
that the workqueue will be freeze before system enter suspend.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>