irqchip: irq-imx-gpcv2: Disable cpuidle if no GPC_CORE_WAKE in EL3
commit 2fe889b77c7cb344996f42bd596d714538858eb9 from
https://source.codeaurora.org/external/imx/linux-imx

Not all EL3 have the FSL_SIP_CONFIG_GPC_CORE_WAKE, therefore disable
the cpuidle to avoid all the cores going to sleep ending up with a
hang. This allows all the EL3 implementations to work with i.MX8MQ
even if they do not support core wake-up through GPC as a workaround.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>