net: stmmac: add Enhanced Tx Descriptor support in main flow
commit 1df51f6ffc148f212c18f36db5a8d83560fb853c from
https://github.com/intel/linux-intel-lts.git

Add support for Enhanced Tx Descriptor in stmmac_main.c for all
Tx related functions.

For TSO, it is available in DWMAC v4.0 and above, so it is not
applicable for extended descriptor which is used in DWMAC v3.5.

The tail pointer for supporting enhanced and extended descriptor
is made to be updated correctly.

Signed-off-by: Kweh Hock Leong <hock.leong.kweh@intel.com>
Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>
Signed-off-by: Tan, Tee Min
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>