net: stmmac: add FPE HW tunables setting
commit 313fe5e7dc3738904029cc57cc99fe05e74e5dd6 from
https://github.com/intel/linux-intel-lts.git

AFSZ is meant to change the non-final fragment size of preempted
frame. A smaller value here means express frame will be transmitted by
express MAC as soon as preempted MAC (pMAC) has collected the configured
fragment size, default to 64-byte fragment.

HADV is meant to specify the time (in nano-seconds) for pMAC to
switch from release to hold state in advance.

RADV is meant to specify the time (in nano-seconds) for pMAC to
switch from hold to release state in advance.

Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>
Signed-off-by: Voon Weifeng <weifeng.voon@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>