ASoC: ti: davinci-mcasp: Support for serializer reconfiguration
commit 4f5f0f375066ca20e35f05e28a54de26bd969edc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Currently the serializer direction can be set only once and it is static.
If the same McASP is connected to different codecs with different sets of
serializers, there is a need to change the serializer pins depending on
which codec is being used currently.

With the new function machine drivers can feed new serial_dir array if
needed.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>