arm64: dts: imx8qxp-mek-dsp: Change compatible string for old DSP driver
commit 6076103a33e88b25cb8a0647140005ef070f2f4d from
https://source.codeaurora.org/external/imx/linux-imx

Currently there are two Linux DSP drivers. We will use the following
compatible strings depending on which driver is used:

1) fsl,imx8qxp-dsp-v1

Used for FSL DSP Linux driver that works with the old audio framework written
completely inside NXP.

2) fsl,imx8qxp-dsp

Used for SOF Linux driver. This will be used in the future, but until
SOF Linux driver / FW will completely support all the features we will
need to support 1) and 2) in parallel.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>