clk: imx7ulp: remove IMX7ULP_CLK_MIPI_PLL clock
commit 17f458d0a4373b8c6d824b343b1b3dad9532c2f3 from
https://source.codeaurora.org/external/imx/linux-imx

The mipi pll clock comes from the MIPI PHY PLL output, so
it should not be a fixed clock.

MIPI PHY PLL is in the MIPI DSI space, and it is used as
the bit clock for transfering the pixel data out and its
output clock is configured according to the display mode.

So it should be used only for MIPI DSI and not be exported
out for other usages.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>