MLK-11340-38 usb: chipidea: imx: enable different wakeup setting
commit 42c246213d9127206ee43e0906ed4d51e4b2c073 from
https://source.codeaurora.org/external/imx/linux-imx

We have different wakeup setting for different roles:
For peripheral-only mode, we may only enable vbus wakeup.
The Micro-AB cable should not be considered as wakeup source.
For host-only mode, the ID change or vbus change should not be
considered as wakeup source. For OTG mode, all wakeup setting
should be considered as wakeup source.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>