mfd: Add audiomix support
commit 5f5a9c97f5917e6659316249222c893663e1b790 from
https://source.codeaurora.org/external/imx/linux-imx

Some of the i.MX SoCs have a IP for interfacing the audio dedicated IPs with
clocks, resets and interrupts, plus some DSP specific control registers.
To allow the functionality to be split between drivers, this MFD driver is
added that has only two purposes: register the devices and map the entire
register addresses. Everything else is left to the dedicated drivers that will
bind to the registered devices.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Leonard Crestez <leonard.crestez@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>