MLK-11293: gpio: pca953x: add device_reset() call
commit 55f45b22e4f84f7b99c9d5ca3e85318301d5e357 from
https://source.codeaurora.org/external/imx/linux-imx

The pca953x type of devices, e.g. max7310, may have a reset which needs
to be handled to get the device start working.  Add a device_reset()
call for that, and defer the probe if the reset controller for that is
not ready yet.

The patch merge below two commits:
1a5c743f463f("MLK-11293: gpio: pca953x: add device_reset() call")
7500d014297a("MLK-13773 gpio: pca953x: correct device_reset() return
	      value check on kernel 4.1")

Signed-off-by: Shawn Guo <shawn.guo@freescale.com>
Signed-off-by: Fugang Duan <fugang.duan@nxp.com>