bus/fsl-mc: Add a new parameter to dprc_scan_objects function
commit e45893bf91e3a5cfe67c2b6e2e2ae343973e0e42 from
https://source.codeaurora.org/external/imx/linux-imx

Prepare the dprc_scan_objects function to be used by
the VFIO mc driver code. The function is used to scan the mc
objects by the bus driver. The same functionality is
needed by the VFIO mc driver, but in this case the
interrupt configuration is delayed until the userspace
configures them. In order to use the same function in both
drivers add a new parameter.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>