dpaa2-eth: Keep congestion group taildrop enabled when PFC on
commit c3b30b61f547606a4fe26842aea17b567d7b10c9 from
https://source.codeaurora.org/external/imx/linux-imx

Leave congestion group taildrop enabled for all traffic classes
when PFC is enabled. Notification threshold is low enough such
that it will be hit first and this also ensures that FQs on
traffic classes which are not PFC enabled won't drain the buffer
pool.

FQ taildrop threshold is kept disabled as long as any form of
flow control is on. Since FQ taildrop works with bytes, not number
of frames, we can't guarantee it will not interfere with the
congestion notification mechanism for all frame sizes.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>