clk: imx8mm/mn/mq: add imx_clk_init_on
commit c52a8c6e9f08e6ebddc123facc712b01cdc4f6cf from
https://source.codeaurora.org/external/imx/linux-imx

When we need to support dual linux with jailhouse, there is no clock
controller in 2nd inmate linux cell, it relys on the first linux to
configure the clock ready and on. So we add those clocks required for
the 2nd linux in dts to make them prepare enabled, and pass
clk_ignore_unused to the 1st linux, then the 1st linux will not gated
off the clocks. So the 2nd linux could use IPs without touching clocks.

Signed-off-by: Peng Fan <peng.fan@nxp.com>