drm/imx/hdp: add colorspace property
commit d1805e3d8f6b19953e9b2339e72e9cbc856522a0 from
https://source.codeaurora.org/external/imx/linux-imx

iMX8MQ has the ability to adjust the DCSS output pipe gamut and nonlinearity
depending on the HDMI connector capability. Userspace can explicitly set this
property if it decides, based on EDID parsing, that the sink supports REC.2020
and it wants to switch when it plays HDR10 content.

Otherwise, the kernel will use the default settings specified in the HDMI
specifications.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>