gpu: imx: dpu: Access regs in display engine units wo holding mutex
commit accaed25e29ef76376aec9ff31ca32b1ac4221eb from
https://source.codeaurora.org/external/imx/linux-imx

We don't need holding mutex when accessing registers in display engine
units, because KMS is the only relevant client driver and it has ww mutex
mechansim to ensure there is no race condition on the CRTC resources.
Also, we are naturally safe when the driver initializes the units at the
probe and system power management stages.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>