gpu: imx: dpu: Access regs in display engine units wo holding mutex
commit 4e25f97a56f23afdfc5e5e9b710f6a44326febb7 from
https://source.codeaurora.org/external/imx/linux-imx

We don't need holding mutex when accessing registers in display engine
units, because KMS is the only relevant client driver and it has ww mutex
mechansim to ensure there is no race condition on the CRTC resources.
Also, we are naturally safe when the driver initializes the units at the
probe and system power management stages.

Signed-off-by: Liu Ying <victor.liu@nxp.com>