drm/imx: dpu: kms: Support full screen CRTC background
commit 174ce27f6961ccc8a34f77f9db1ffcd51931a099 from
https://source.codeaurora.org/external/imx/linux-imx

The CRTC background should be full screen instead of partial
screen, because the DRM core is likely to add configurable
background color support in the future.  We may cover the full
screen with ConstFrame0/1, upon which builds planes.  With this,
it is easier to compute each plane's layer offset vs CRTC start
point and all ConstFrame units can be controlled by CRTC.

Signed-off-by: Liu Ying <victor.liu@nxp.com>