MLK-22951 mxc: vpu_malone: add missing semicolon
commit 6dadcfd56c4836c9bb0d4056ec34bf2526c3ec95 from
https://source.codeaurora.org/external/imx/linux-imx

Add a missing semicolon. This missing semicolon is not breaking the
build in defconfig since the next MODULE_DEVICE_TABLE() expand to an
empty string, so the next semicolon is used. When this macro expands
to a non-empty string, the build breaks. This was found by a
"make allmodconfig" build.

Signed-off-by: Julien Olivain <julien.olivain@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>
Acked-by: Zhou Peng <eagle.zhou@nxp.com>