MLK-22952 mxc: vpu_windsor: add missing semicolon
commit fe8af647091a1665d1b2ef613f9a8e4d3e77a15f from
https://source.codeaurora.org/external/imx/linux-imx

Add a missing semicolon. This missing semicolon is not breaking the
build in defconfig since the next MODULE_DEVICE_TABLE() expand to an
empty string, so the next semicolon is used. When this macro expands
to a non-empty string, the build breaks. This was found by a
"make allmodconfig" build.

Signed-off-by: Julien Olivain <julien.olivain@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>
Acked-by: Zhou Peng <eagle.zhou@nxp.com>