bus: arm-ccn: add a new driver managing the ARM CCN interrupts
commit  e4ee41c17b3b119175894e1eceef1b1e34244aaf from
https://github.com/axxia/linux-yocto.git axxia-linux/v5.10/standard/axxia-dev/base

It happens out there CCN PMU is not the only driver of ARM CCN.
Among them the EDAC L3 detecting and reporting the HW errors is.
With the ARM CCN resources being shared, namely the control registers
but most importnatly an interrupt line (INTREQ), a need arose to
separate the interrupt handling routine and assign it to a ARM CCN core
driver managing and dispatching the interrupts to its clients, if any.

In detail, ARM CCN signals errors and PMU event counter overflow through
the INTREQ. If a client driver (error or PMU) exists it shall register its
interrupt service routine with the core. The core then dispatches
the interrupts to the clients respectively. If neither of the clients exist
it simply reports the error/s and clears the interrupt line preparing
the ARM CCN interrupt processing for arrival of the next interrupt.

This commit does two things: takes out the interrupt processing off CCN
PMU and creates the interrupt core driver making CCN PMU use the core.

Signed-off-by: Marek Bykowski <marek.bykowski@intel.com>
Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>