LF-58: sata: imx8: fix the calibration failure of phy
commit 7ef83a3b25cd3c1b2962cc5b0bcc13526011e393 from
https://source.codeaurora.org/external/imx/linux-imx

The calibration of the SATA PHYX1 is derived from the PHYX2.
Add the release of the PHYX2 APB reset, and turn on the HW gated
pipe_pclks of PHYX2 to make sure that the calibration of SATA PHYX1
can be finished successfully.
Adjust the APB reset of SATA, it should be released firstly.
The EPCS configrations should be placed before the EPCS reset.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>