drm/bridge: adv7511: Fix of_reconfig
commit c8b30c59051788225ae703bcf78bd8420aa7fdea from
https://source.codeaurora.org/external/imx/linux-imx

Currently, if the actual i2c device chip (ADV75**) is not found, the
remote node is disabled completely. This causes undefined behavior, if
that device was probed and being used.
In order to fix this behavior, don't disable the remote device, just
detach the remote's endpoint, so that the remote device will just
abandon the connection.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>