enetc: Set MDIO_CFG_HOLD to the recommended value of 2
commit 73ea4213a887c2fdd0082e2b26d4af4a2b5219bb from
https://source.codeaurora.org/external/imx/linux-imx

This increases the MDIO hold time to 5 enet_clk cycles from the previous
value of 0. This is actually the out-of-reset value, that the driver was
previously overwriting with 0. Zero worked for the external MDIO, but
breaks communication with the internal MDIO buses on which the PCS of
ENETC SI's and Felix switch are found.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
(cherry picked from commit 4daae5b2f19a58ae29a30172af4f2e454d8dc3c5)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>