MLK-17077 usb: typec: clear vbus change event in irq handler
commit bf3105a36beec5a31b3aef3e2a10d4917adb1b06 from
https://source.codeaurora.org/external/imx/linux-imx

For vbus change event, we need read the vbus status to clear
the alert. Current code do this in queue work, this has problem
on single core running, the queue work of vbus change may have
no chance to be scheduled as we continue receive the vbus change
event in threaded irq.

Acked-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>