MLK-17077 usb: typec: clear vbus change event in irq handler
commit 2474e149bc1fc6978730fe5a62ed11325010c47f from
https://source.codeaurora.org/external/imx/linux-imx

For vbus change event, we need read the vbus status to clear
the alert. Current code do this in queue work, this has problem
on single core running, the queue work of vbus change may have
no chance to be scheduled as we continue receive the vbus change
event in threaded irq.

Acked-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>