LF-425-2 drm/bridge: sec-dsim: use 'attach-bridge' for bridge attach
commit 2dc660c20e069a4d4adb424dca1be07585b8c65f from
https://source.codeaurora.org/external/imx/linux-imx

Since the port nodes under dsim device can be detached
dynamically by other devices. So use the 'attach-bridge'
property to imply that if there is any available bridge
endpoint node exists. If not, the panel node existence
is required to be detected. If still not, whole display
components binding should be ended which can avoid the
endless defer probe problem when no available display
devices can be attached to dsim.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>