spi: pl022: Add enable-dma processing
commit  38df74b7b7702635ba8ececdad7c690c68411105 from
https://github.com/axxia/linux-yocto.git axxia-linux/v5.10/standard/axxia-dev/base

The pl022 driver defaulted to DMA enabled. The new code will
try to read the of_property "enable-dma".
If the property is read correctly, the value will be used, if
the property is not available or is failed to be read the
enable-dma property will be set to the default, ON.

Signed-off-by: Charlie Paul <charlie.paul@windriver.com>
Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>