Fixed qbman alignment error in the virtualization context
commit 4bb056f31e52100dcb4b3439ec88ecdad553d8c1 from
https://source.codeaurora.org/external/imx/linux-imx

When running as a guest, under KVM, the CENA region is mapped
as device memory, so uncacheable. When the memory is mapped
as device memory, the unaligned accesses are not allowed.
Memcpy is optimized to transfer 8 bytes at a time regardless
of the start address and might cause alignment issues.

Signed-off-by: Diana Craciun <diana.craciun@nxp.com>