MLK-23252-4 gpu: imx: lcdifv3: Add compatible string for LDB's LCDIFv3
commit 199fc5da2e05e4789f26b9c7e61ccb1c75315375 from
https://source.codeaurora.org/external/imx/linux-imx

The current lcdifv3 common driver uses of_device_id data to determine
the data enable signal polarity of LCDIFv3.  As required by LDB, the
polarity has to be "de_invert", so this patch introduces a compatible
string for LDB's LCDIFv3 and sets the "de_invert" flag to true in the
relevant of_device_id data.

Reviewed-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>