MLK-23181-3 usb: cdns3: gadget: handle multiple trb request properly
commit f9e2ca657ff28801410a16aad9cd479b14f85ec4 from
https://source.codeaurora.org/external/imx/linux-imx

We use polling EP_TRADDR to judge if the trb has finished, so for
request which includes multiple trb, we may quit the trb judgement
in the middle of request, in that case, we can't giveback the
request. We should only giveback the request when all the TRB in
this request has finished.

Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>