MLK-23287-1: ASoC: fsl_aud2htx: Add aud2htx module driver
commit 7fcdf7d7e33f0af27470cc39569f5948ad5ba84c from
https://source.codeaurora.org/external/imx/linux-imx

The AUD2HTX is a digital module that provides a bridge between
the Audio Subsystem and the HDMI RTX Subsystem. This module
includes intermediate storage to queue SDMA transactions prior
to being synchronized and passed to the HDMI RTX Subsystem over
the Audio Link.

The AUD2HTX contains a DMA request routed to the SDMA module.
This DMA request is controlled based on the watermark level in
the 32-entry sample buffer.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>