MLK-23394-1: dmxdev: return fail if ringbuffer overflow.
commit 92490d18374b2b4e32c716343b93226283119141 from
https://source.codeaurora.org/external/imx/linux-imx

Ring buffer will flush if set the error when overflow. The ring
buffer in our case is always full as our source is from file,
not tuner. So we need disable flush and send data to ring buffer
again if ring buffer full. Please revert the patch if use real
tuner.

Signed-off-by: Bing Song <bing.song@nxp.com>
Reviewed-by: Jian Li <jian.li@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>