sdk_dpaa: ceetm: fix recursive dependencies
commit ebe781d9ed24b92ca2c983f030e95bbfbd312785 from
https://source.codeaurora.org/external/imx/linux-imx

Due to dependencies between the fsl_ceetm and fsl_dpa modules, remove
the module support for the ceetm driver and integrate it into the main
DPAA Ethernet driver.

The registration of the CEETM Qdisc was the only operation done at
module init. Pass the management of the Qdisc register and unregister
operations to the loading and unloading of the DPAA driver.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>