tty: pl011: Add support for configurable wordlength
commit  4c74f64859c921c39693ad31c30c6ad5892a8675 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Currently we are not writing to the line control register
to update the wordlength.
Add support for configurable word length by updating the lcr.

Currently if we echo characters to the non-console uart the
no characters are displayed. The patch fixes the same.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>