v4l: xilinx: sdirxss: Add support to modify GT clock
commit  d6a15ca14d330e1bb08ee951734c2c6aeec77f18 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

In 12G mode, the GT clock should be modified to integer or fractional
clock based on the mode. The clock applied to GT is 148.5MHz in case of
integer frame rate mode (e.g. 12G 3840x2160p60). The clock should be
148.35MHz in case of fractional frame rate mode (e.g. 12G
3840x2160p59.94).
The irq handler is changed to threaded type as now we use CCF to modify
the clock in the irq handler.
The driver found print in the probe() is moved to the end of probe().

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>