MLK-23234-4 gpu: imx: add LCDIFv3 core driver
commit f9cac90767fbeeeb163a1ea999e567ff6103ce77 from
https://source.codeaurora.org/external/imx/linux-imx

The LCDIFv3 core driver is responsible to provide
controller registers configuration and create the
platform devices for the child port nodes. And the
platform devices later will attach to the related
DRM/KMS drivers via name match. And the LCDIFv3 is
completely different from the LCDIF controller
which is used on imx8mm and imx8mn platforms.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>