arm64: dts: ti: k3-j721e-main: Add ICSSG MDIO nodes
commit 875e0bdda32e73e9291e3eff65a7ed03ef6362a1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The ICSSGs on K3 J721E SoCs contain an MDIO controller that can
be used to control external PHYs associated with the Industrial
Ethernet peripherals within each ICSSG instance. The MDIO module
used within the ICSSG is similar to the MDIO Controller used
in TI Davinci SoCs. A bus frequency of 1 MHz is chosen for the
MDIO operations.

The nodes are added to the common k3-j721e-main.dtsi file and are
disabled. These need to be enabled in the respective board files
supporting J721E SoCs and where the ethernet is pinned out and
connected properly.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>