staging: vchiq_arm: Give vchiq children DT nodes
commit  1111c967b75761d168d7d5f40a3736260ddc783a from
https://github.com/raspberrypi/linux.git rpi-5.10.y

vchiq kernel clients are now instantiated as platform drivers rather
than using DT, but the children of the vchiq interface may still
benefit from access to DT properties. Give them the option of a
a sub-node of the vchiq parent for configuration and to allow
them to be disabled.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>