HACK: net: netcp: ale: add sysfs interface for control, ale table show
commit 19d97b8beb0e291d80a1afdfcd44d85d83aa04ac from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This patch enhances the ale driver to support sysfs interface for control
and ale table show. This is done only for ALE hardware found on NetCP
by reading the ale version and register the interface only for ALE on
NetCP hardware.

This patch is being marked as a hack because these controls should be
implemented using DSA/Switchdev support in the driver. This requires
the driver to be migrated to  that framework. Also, some other controls
require additional enhancements to user space tools such as iproute and
ethtool.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>