LF-908-1: fsl_ssi: fix noise issue for master and mono mode
commit 92601d2f59da3bdc173e3a01689b056b4ce41187 from
https://source.codeaurora.org/external/imx/linux-imx

The driver will switch to normal mode for 1 channel, so the slot width
is not fixed to 32 bit in this mode, and even in mono mode, the slot
number is 2.
But if the bit clock calulation still use the bit width=32, and slot
number=1, then it cause the output frameclock is not correct. So there
will be noise in output sound.

This patch is to change the bit clock formula of mono mode to be
2channels * params_width * sample rate.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>