mmc: sdhci-of-arasan: Modified SD default speed to 19MHz for ZynqMP
commit  6009536babad4027024222a5b2e0a7b3dada86f2 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

SD standard speed timing was met only at 19MHz and not 25 MHz, that's
why changing driver to 19MHz. The reason for this is when a level shifter
is used on the board, timing was met for standard speed only at 19MHz.
Since this level shifter is commonly required for high speed modes,
the driver is modified to use standard speed of 19Mhz.

Signed-off-by: Manish Narani <manish.narani@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>