xilinx: Add a github template telling people to not use pull requests there
commit  5a6fcd95636213bc16f59e717b91b2b17292c752 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

A lot of people are trying to send pull requests via github but this is not
the right channel to use. Linux project is doing review over emails that's
why use git@xilinx.com if you want to contribute changes to this
repository.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>