Revert "net: phy: call state machine synchronously in phy_stop"
commit  975a6a7b7c1760f9d3f579d07058d91d9e5ee3bd from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This reverts commit e8cfd9d6c7727a067b38dbe7655ca02377fdb301.

An issue was observed with TI DP83867 PHYs on Xilixn boards where ifconfig
down/up does not work. This was traced to changes in phy_stop where the
state machine is stopped. It is as yet unclear why this is affecting this
PHY and it is being debugged. These patches are being reverted as a
workaround because this affects basic functionality.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>