LF-941: ASoC: fsl_easrc: Switch to 32 taps for resampling filter
commit dc5a4818433a59c187fa52b7d4585940ba6b2da5 from
https://source.codeaurora.org/external/imx/linux-imx

With 128 taps of resampling filter, there is performance issue
for high sample rate and mulitchannel. For example, there is
noise issue for converting 48kHz to 768kHz, channel number
is larger than 5 channels, and p2p case.

So in order to increase performance, we reduce the taps number
to 32 in default, the quality is relatively downgraded.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>