arm64: dts: ti: k3-am654-evm-tc358876.dtso: fix clock node name
commit 0a0e8aca1be0bda2e57374d1e0e28fb90ef2abcf from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The fixed clock for tc358876 is named 'clock' which is a very generic
name and easily leads to conflicts. I encountered this with a similarly
named fixed-clock node in k3-am654-base-board.dts, causing OV5640 not to
work when tc358876 overlay had been loaded.

Rename the node to 'fixed-clock-tc358767'.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>