MLK-23560-2: ASoC: SOF: Enable PM support
commit b81f048f9d0c65dcaa87c51929fbd682246e7862 from
https://source.codeaurora.org/external/imx/linux-imx

We need to provide PM macros in order for the SOF core
to suspend/resume the DSP.

PM support on the DSP side is quite limited and we are mostly
doing PM for ARM core.

Without this patch DSP won't be restarted after a resume causing
SOF driver calls to time out.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>