SSI-87: firmware: imx: scu: Support reception of any size message
commit 16338e4fe9b03a812b4ab659992b614dc10c9bcc from
https://source.codeaurora.org/external/imx/linux-imx

The implementation was limiting the size of a message which can be
received to 4 but soem response can be bigger. For example the
response of the 'sc_seco_secvio_config' API is 6 words.

This patch removes this limitation relying on the count of word
received instead of the index of the chan.
It does so by duplicating imx_scu_call_rpc as imx_scu_call_big_rpc
in order to cahnge the RX method using imx_scu_big_rx_callback
instead of imx_scu_rx_callback.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>