HSD #1507870299: video: fbdev: fix dma allocation error
commit  ed474fd6c3a241adbcea56508087d82f0e795f62 from
https://github.com/altera-opensource/linux-socfpga.git

A NULL pointer is being passed into the dma_alloc_coherent as a 'dev'
pointer, which causes a kernel panic as it tries to dereference a NULL
pointer. This patch fixes the problem by passing a device into the
function. The same is applied for dma_free_coherent.

Signed-off-by: Ooi, Joyce <joyce.ooi@intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>