MLK-23754-22 remoteproc: imx_rproc: add i.MX8QXP CM4 start/stop with power domain support
commit f960ed8a9f45402ce47335ba5acf348275be1225 from
https://source.codeaurora.org/external/imx/linux-imx

When CM4 image needs to be loaded from Linux userspace, we need
power on CM4 CORE, MU_1A and etc. Then start CM4 CORE.

Because we need to add multilple power domains, default power
domain attach probe only support one power domain,
we get the power domain in probe and mark on.

We are lucky that the TCML space could be accessed even without
power on the resources, so it is ok power on in probe. If not,
elf load will not be able to load segments to TCML.

Reviewed-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>