HSM-267-3: mailbox: imx: Add timeout based on jiffies
commit eea787c67cd796fa5a05b06d440a68e796ab3c8a from
https://source.codeaurora.org/external/imx/linux-imx

The remote processor can take some time to read/write to
the MU and the previous timeout was not based on a temporal
value. It is better to use a temporal variable so the code
will always have the same behavior.

There is no specification about the timeout to set, just to have
something in the range of few seconds, so chosen 3.

This patch changes the timeout to be based on time.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Horia Geantă <horia.geanta@nxp.com>