remoteproc: zynq: Add explicit nodes for vrings
commit  688246a89c54b67ae216777ade0089ca8cb4f482 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch changes zynq driver to specify the firmware memory that
is required explicitly. Without this, the driver will not work, as the CMA
allocated memory will instead be used.

In conjunction this patch also enables user to specify the vring and
shared buffer location to ensure that only this aforementioned
memory space is used.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>
Acked-by: Wendy Liang <wendy.liang@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>