MLK-23635: fw: imx: seco_mu: Fix messages at boot
commit 87bcafe6925d24bcd75b16aa71c4f2aecb13daf8 from
https://source.codeaurora.org/external/imx/linux-imx

At boot, the driver can print error messages because it has been
deferred so they should not be printed.
The driver also prints the version of the SECO which is not useful.

This patch fixes the print of messages due to defer and remove the
print of the SECO version.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Horia Geantă <horia.geanta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>