MLK-23792-1: ASoC: fsl_sai: Monitor spdif rx clock in imx8mm
commit 02d362f8d6dac7bd2357edf30420db8648860c77 from
https://source.codeaurora.org/external/imx/linux-imx

As we use one spare sai instance to monitor the spdif rx
clock,  there isn't belong to a sound card, we can't access
the registers by amixer controls.

So remove the amixer controls, replace them with the device
attribute.

And add an additional device attribute for enablement of
monitorring spdif. This feature only be supported on imx8mm.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>