MLK-23792-1: ASoC: fsl_sai: Monitor spdif rx clock in imx8mm
commit 10bcbef24ba7cf95a199830890d243bce97f33be from
https://source.codeaurora.org/external/imx/linux-imx

As we use one spare sai instance to monitor the spdif rx
clock,  there isn't belong to a sound card, we can't access
the registers by amixer controls.

So remove the amixer controls, replace them with the device
attribute.

And add an additional device attribute for enablement of
monitorring spdif. This feature only be supported on imx8mm.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>