PCI: layerscape: Change to use the DWC common link-up check function
commit 793987dd77ebca9868fc7f1d6c433536b6825eb9 from
https://source.codeaurora.org/external/imx/linux-imx

The current Layerscape PCIe driver directly uses the physical layer
LTSSM code to check the link-up state, which treats the > L0 states
as link-up. This is not correct, since there is not explicit map
between link-up state and LTSSM. So this patch changes to use the
DWC common link-up check function.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>