phy: cadence: cadence-sierra: Don't configure if any plls are already locked
commit 2f3f703a7db7603433b97b2be1e5ad3897b2a61b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Serdes lanes might be shared between multiple cores in some usecases
and its not possible to lock PLLs for both the lanes independently
by the two cores. This requires a bootloader to configure both the
lanes at early boot time.

To handle this case, skip all configuration if any of the plls are
already locked. This is done by adding an already_configured flag
and using it to gate every register access as well as any phy_ops.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>