MLK-23913-4 usb: cdns3: gadget: set fast access bit
commit 62faac59a9cf9ffc10bcb90ba2e454aadd388b59 from
https://source.codeaurora.org/external/imx/linux-imx

Below is the recommendation from Cadence designer:
Using this bit to be sure that PHY clock is keeping up in active
state. It's good to keep Fast Access bit enabled as long as there
is any access to USB register.

It is used to fix the potential ARM core hang when visit controller
register after DEVDS is set.

Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>