MLK-23885 drm/imx: ldb: Add mode fixup for i.MX8mp LDB in ->atomic_check()
commit c09f1acd72dcb5bdbd04cc0adf11f3400ee56ef4 from
https://source.codeaurora.org/external/imx/linux-imx

Due to limited video PLL frequency points on i.MX8mp, the exact pixel clock
rate of a panel's video mode is likely unsupported by the clock tree.
The driver would take a slightly fixed up pixel clock rate to match with
the clock tree capability and the panel should still work with that.
This should work based on the agreement that the chosen panel should work
with the existing PLL frequency points.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Fancy Fang <chen.fang@nxp.com>