arm64: dts: imx8mp: Make audio_blk_ctrl single node for combo driver
commit 482838aa8bf0ad44dbb8b0eeb83d1dd7904f8c03 from
https://source.codeaurora.org/external/imx/linux-imx

Make the audio_blk_ctrl dts node a single one and change all the references
to it accordingly. Do this in order to have a single combo driver for
audio_blk_ctrl. Some of the features of the audio_ctrl will be used by some
different drivers in a way those drivers will know best, so adding the
syscon compatible we allow those to do just that. Those drivers will
have to also add the audiomix PD to their devicetree node in order
to make sure audio_blk_ctrl is on while they are operating.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Tested-by: Daniel Baluta <daniel.baluta@nxp.com>