MLK-23906-2: ASoC: fsl_aud2htx: Switch to imx-pcm-dma-v2
commit 7cdd916faab4c7623db1c7c8d2b2214ab500c2b8 from
https://source.codeaurora.org/external/imx/linux-imx

With imx-pcm-dma, the dma channel is created in probe, when
there is power domain attached with dma device, the power
domain will be enabled when channel is created, then the
power of the domain will be always enabled from beginning.

So switch to imx-pcm-dma-v2, then the dma channel will be
created when playback or capture really started.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>