MLK-24002-1 drm/imx: lcdifv3: enable shadow load for plane update
commit e532a1d92c8a3cca705526111030cab324c2b84d from
https://source.codeaurora.org/external/imx/linux-imx

After the atomic plane update, the shadow load should be
enabled to make sure its update can take effect on next
frame in any cases. And this enable is better to be done
in CRTC's atomic_flush() which is called after plane's
atomic_update() is called.

Besides, the shadow load enable in controller enable is
unnecessary, so remove it.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>