LF-1373-7: drm/bridge: nwl-dsi: Add support for 8qm & 8qxp
commit 68fa4eff3c2fa5597984204fced6be9bdfb55008 from
https://source.codeaurora.org/external/imx/linux-imx

The current nwl-dsi driver was initially written for 8mq, but the same
IP can also be found in 8qm and 8qxp SoCs.
The changes needed for 8qm and 8qxp are regarding to the architecture
differences:
- on 8qm and 8qxp, DSI reset can be accessed only through system
controller.
- also, on 8qm and 8qxp there are two nwl-dsi IP blocks, instead on just
  one, as found in 8mq
- on 8qxp, the dphy is a shared phy with the lvds IP block, so we need
  to account this too, when controlling the mipi dphy.

This patch addresses the above and fixes different problems that occured
with the addition of these new features.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>