LF-1405-1: [8QM_MEK/8QXP_MEK]mxc:vpu_malone: support insmod separately
commit  from
https://source.codeaurora.org/external/imx/linux-imx

delay the creation of video device node in order to avoid inmod failure

in probe function, if create video node in a early stage,
the video node may be created before initialize the vpu dev,
and in such case, open video node,
the mu mailbox may be not requested,
then it will led to download firmware fail.

so create the video device node at the end of the probe function.

Signed-off-by: Ming Qian <ming.qian@nxp.com>