drm/vc4: Adopt the dma configuration from the HVS or V3D component
commit  eecf2e17e7e9c6fa236e9d4a699bc90b57fe7d81 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

vc4_drv isn't necessarily under the /soc node in DT as it is a
virtual device, but it is the one that does the allocations.
The DMA addresses are consumed by primarily the HVS or V3D, and
those require VideoCore cache alias address mapping, and so will be
under /soc.

During probe find the a suitable device node for HVS or V3D,
and adopt the DMA configuration of that node.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>