HACK: dma-buf: heaps: carveout: Add support for cached carveout heaps
Remove the "no-map" property from the carveout DT node to have kernel
and userspace mappings cached.

This is currently a hack as I have not yet added a good way to clean
these buffers out of the CPU cache. These should only be used on
io-coherent platforms such as K3.

Signed-off-by: Andrew F. Davis <afd@ti.com>
[Xulin: Original patch taken from TI PROCESSOR-SDK-LINUX-J721E_08.00.00.08]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>