genirq/cpuhotplug: axxia: Enable the force flag
commit  4bee54b43f6adb26308a11837ae48975cab960bc from
https://github.com/axxia/linux-yocto.git v5.10/standard/preempt-rt/axxia-dev/base

The force flag was not used for X86 but is required by ARM.
This changes adds the force flag to the AXXIA AXM55xx chips.

Setting the flag in the CPU hotplug code causes the cpu online masking to
be blocked on these irq controllers and results in potentially affining an
interrupt to the CPU which is unplugged, i.e. instead of moving it away,
it's just reassigned to it. The recommendation for this change originated
from Thomax Gleixner at Linutronix.

Signed-off-by: Charlie Paul <charlie.paul@windriver.com>
Signed-off-by: John Jacques <john.jacques@intel.com>
Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>