LF-1599 media: platform: mxc_vout: Do not check 's->r.width/height<0' in ->vidioc_s_selection
commit 6d82f510f98a1947740355749f2551d779c8a2f3 from
https://source.codeaurora.org/external/imx/linux-imx

The commit f90580ca0133 ("[media] videodev2: Set vb2_rect's width
and height as unsigned") changes the width and height fields of
struct v4l2_rect from type __s32 to type __u32, which makes Coverity
complain that it's unnecessary for the platform driver to check on
the fields to ensure they are not less than zero.  This patch removes
the check to make Coverity happy.

This fixes Coverity issue: CID 9176743.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Sandor Yu <sandor.yu@nxp.com>