net: ti: prueth: dual_emac: Add support for timestamping tx packets
commit 4d8bd9f0cecaaafb71d8da0e25f743689bcbc635 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When PTP support is enabled in dual emac firmware, the firmware timestamps
the ptp packets that are going out via prueth ports. The timestamp value
is copied to a shared memory location. After timestamping, a host irq is
raised by the firmware as a notification to read the timestamp.

Add support for handling this ptp tx irq, read the tx timestamp and pass
it to userspace layer.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>