ASoC: SOF: imx: Add code to manage DSP related clocks
commit a8dec59321c6d5a61d46b717cf61db347c22881a from
https://source.codeaurora.org/external/imx/linux-imx

We introduce arrays of clocks that will be enabled/disable
at resume/suspend in order to save power.

There are two types of clocks:
	* DSP IP clocks
	* DAI clocks

Normally, DAI clocks would be managed by the DSP firmware
but this is hard to achieve now, so just manage them from
Application Processor side for now.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>