tty: serial: linflexuart: Resume DMA transfer after polling mode finished
commit 33fd2ac3d9ae193acdba0e15febbd981cab10d65 from
https://source.codeaurora.org/external/autobsps32/linux

When there are both kernel and userspace output, the
last one can't be seen after a while. The userspace console hangs until
the user wakes up it by sending a character.
The fix is to resume the tty output and the DMA transfer after polling
mode output is finished.

Issue: ALB-5329
Signed-off-by: Js Ha <js.ha@nxp.com>
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>