misc: xilinx-ai-engine: Add function to enquire if a partition is available
commit  4dc64a65bf4458d4946b55bbd78e7f75980ca718 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Add a function which is available to other kernel module to enquire if a
partition is available. We have request and release partition functions
available, however, the release function is to reduce the reference count,
it is not a synchronous call to release the resource of the partition.
And thus, it needs a function to check if a partition is actually
released.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>