misc: xilinx-ai-engine: use bitmask for AI engine tile type
commit  3ba210f0203751c97724a75e65c083915f885e26 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This is to fix the register type check in register verification. There
are some registers available in more than one type of tiles. Such as
registers available to SHIM PL are also available to SHIM NOC. And thus,
bits mask for AIE tile types is easier to indicate the registers are
available in which types of tiles.

This information is used to validate registers access request from users
to see if the registers should not be configured from user of a tile.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>