dt-bindings: clock: Added documentation for global STM time source
commit 49adbdd588c8e85a92913b756b2d43de542fe485 from
https://source.codeaurora.org/external/autobsps32/linux

The documentation file explains that the particular usecase for the STM is
as a global time source for the entire SoC. A compatible device should
contain the base address for this peripheral, alongside reference to the
name of the clock groups (stm).

Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Issue: ALB-5093

dt-bindings: clock: Convert NXP stm global timer to json-schema

Convert the NXP stm global timer binding to DT schema format using json-schema.

Signed-off-by: Andrei Botila <andrei.botila@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>