s32gen1: stm: Fix STM not appearing in menuconfig.
commit  2285a37d457bb6afd0232e96a7ece6cdff7fac43 from
https://source.codeaurora.org/external/autobsps32/linux

The STM driver was not appearing in menuconfig because it laked a proper
description in the kconfig file. Modified the Kconfig file so that a
string is provided to be used by menuconfig in order to be have the
configuration option shown.

Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Issue: ALB-5093
Signed-off-by: Meng Li <Meng.Li@windriver.com>