MLK-24937-4 pci: modify AMBA_ORDERING_CTRL_OFF register for rc/ep to avoid vop iperf3 test hang
commit 21b685bedccb3c72e053ac4c9ade05a3ddd58ecb from
https://source.codeaurora.org/external/imx/linux-imx

We met system hang when doing vop iperf3 test, the root cause is there
is a deadlock when non-posted reads can’t be completed before posted
write completion for the PCIe ordering rules.

Here is just a workaround to set the AMBA_ORDERING_CTRL_OFF register to
disable ordering rules on AXI bridge.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>