MLK-24612-1: ASoC: fsl_esai: Add esai client driver
commit d5986dc1d41899a11bd636fe337ef7d503183683 from
https://source.codeaurora.org/external/imx/linux-imx

ESAI client driver is working as cpu dai in front-end.
ESAI is an unity, which can't be separated to multiple device
in hardware. so we need to separate it in software view.

Each client device can accept data from user space indepedently.
Then the data of different client is mixed together before sent
to ESAI interface.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Peng Zhang <peng.zhang_8@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>