iio: versal-sysmon: add driver for Versal Sysmon
commit  7d59f04a059ad72417fef2778da04669688fe279 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Sysmon Driver uses Linux IIO framework, which was used to abstract the
supply voltages and temperatures across the chip as Voltage and Temperature
Channels in the framework. Since there are only 160 supply voltage
registers and 184 measurement points, there is no constant mapping of
supply voltage registers and the measurement points. User has to select
the voltages to monitor in design tool. Depending on the selection, a
voltage supply gets mapped to one of the supply registers. So, this mapping
information is provided to the driver via a device tree. Depending on the
number of supplies enabled in the design, the device tree will contain the
information of name of the supply enabled and the supply register it maps
to.

Signed-off-by: Anand Ashok Dumbre <anand.ashok.dumbre@xilinx.com>
Reviewed-by: Dragan Cvetic <dragan.cvetic@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>