s32gen1: stm-global: Fixed a bug where the global timestamp was not built.
commit  4bfe4a05691ad6c4706d15149b405d621e3a7688 from
https://source.codeaurora.org/external/autobsps32/linux

Due to the fact that the Makefile inside drivers/clocksource entry
coresponding to the universal timesource driver was not prefixed with FSL_
it was not compiling. This commit simply adds the required prefix so that
the driver is automatically generated.

Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Issue: ALB-5464
Signed-off-by: Meng Li <Meng.Li@windriver.com>