misc: xilinx-ai-engine: add tiles in use bitmap
commit  3eafdacc9739898dc133eff8defca157aac0effe from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

As the clock state bitmap itself is not enough to indicate what tiles
are in use. In AIE architecture, to enable a tile, all the tiles below
need to be enabled. When releasing tiles, some tiles are not in use
cannot be released because clock state bitmap was not tracking the
requested tiles. And thus, add another bitmap to track the tiles in
use.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>