drivers: base: export next_driver
commit b8ef4c9afb38a20194fe5d204e835f26d1a374d2 from
https://source.codeaurora.org/external/imx/linux-imx

Exporting the next_driver() function is necessary, for the moment, so
that the dpaa2_mac driver can be forcefully unbound from the DPMAC
object when a kernel driver DPNI is connected to it.
This will be removed once the MC sends interrupts notifying any link
partnet change to all connectable objects (not just DPNIs).

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>