MLK-24862-1 ARM64: dts: imx8mq-evk-dcss-rm67199: add goodix touch support
commit bd81511b80b88476ec67ddf9d8ec687ddb998727 from
https://source.codeaurora.org/external/imx/linux-imx

The new OLED not only use a new DSI panel(Raydium RM67199), but
also use a new touch chip, goodix gt1151q. This touch chip also
share the reset pin with the panel. According to the test, this
touch chip is not that critical with the reset operation, so
let panel control the reset pin.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>