MLK-24862-2 input: touch: goodix: force set the IRQ_TYPE_EDGE_FALLING for GT1151Q
commit e39120247668e3efceccb0775ea352d9421b8319 from
https://source.codeaurora.org/external/imx/linux-imx

For the touch chip use on the new OLED panel(rm67199), when touch
the screen, GT1151Q sends a falling edge pulse via INT pin.

For code logic, it will read the touch internal config data, the bit[1~0]
of config[6] means the interrupt trigger method. For this GT1151Q, this
bit[1~0] is 3 in default, which means IRQ_TYPE_LEVEL_HIGH, which not
match with the datasheet and the hardware behavior. Since the default
level of the INT pin is high, so after driver probe, a lot of useless
touch interrupts keep triggered.
To fix this, need to update the touch firmware. Set the bit[1~0] of config[6]
to value 1. Since we lack the touch firmware file, as a workaround, add a
property "edge-failling-trigger" in dts file, if has this property, force
set the IRQ_TYPE_EDGE_FALLING.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>