spi: spi-cadence-quadspi: Pull TX start and end boundaries from DT
commit 0adfe563dd412778a1747bc18d6a193e9d0f77b9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The TX window from 16 to 48 is not always the best place to look for
rxlow and rxhigh values. Sometimes it might fall outside the range of
valid TX values. So to account for that a mechanism was introduced to
iterate over the TX boundaries to find the correct TX window.

This mechanism makes the algorithm more robust, and it will fail less
often. But the downside is that it would take longer to finish if TX =
16 or TX = 48 are outside the valid TX range. This can be mitigated by
specifying the initial values of the TX boundaries via device tree. This
way, the algorithm is kept robust because it will still iterate over the
TX range if the initial values do not work. But at the same time the
tuning time can be optimized by properly setting the initial boundary
values observed to be stable during testing.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>