staging: xlnxsync: Add new ioctl to set interrupt mask
commit  3877897d2aa6f8e19192778cfffdf99280cb3796 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

New SyncIP RTL supports selective masking of interrupts which was not
working previously. So it's possible now for application to selectively
mask interrupts which it is not using to make any decisions or gain
any information.

For this New IOCTL XLNXSYNC_CHAN_SET_INTR_MASK is implemented along with
interrupt context structure using which application can pass to driver
the interrupts which it doesn't require and thus need to be masked.

So for e.g. if an application doen't use frame done event from syncip
driver then it can mask the corresponding interrupts using this ioctl
thus reducing the overall system overhead and maximizing the
performance.

Also Updated the IOCTL version as API changes.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>