coresight: pmu: Create symlink to the sink device
commit 8077719874c5a4c508091a2ffc50b629131fd59a from
git@git.assembla.com:cavium/WindRiver.linux.git

Create a symlink to the actual sink device from the
ETM pmu device. This is would help the perf tool to
determine features provided by the sink.

Example:
While decoding the trace using perf tool, we need to
know if the sink supports formatted trace. To get this
information, we need to read the FFSR register of the
"enabled sink" which is exposed by sysfs.

Similarly based on the use case we can determine more
properties of the sink if its exposed to ETM_PMU perf
device.

Change-Id: I05a4eba4f7a234cc9c4fde45059dea38dd881266
Signed-off-by: Tanmay Jagdale <tanmay@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>