staging: bcm2835-codec: Allow decode res changed before STREAMON(CAPTURE)
commit  7792edd663040e7751375a36b16092b6aea28299 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The V4L2 stateful video decoder API requires that you can STREAMON
on only the OUTPUT queue, feed in buffers, and wait for the
SOURCE_CHANGE event.
This requires that we enable the MMAL output port at the same time
as the input port, because the output port is the one that creates
the SOURCE_CHANGED event.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>