LF-2321-10: drm/mxsfb: Add support for horizontal stride
commit 58f201d70508f979779f0c789f2cd17404bca7c1 from
https://source.codeaurora.org/external/imx/linux-imx

Besides the eLCDIF block, there is another IP block, used in the
past for EPDC panels. Since the iMX.8mq doesn't have an EPDC connector,
this block is not documented, but we can use it to do additional
operations on the frame buffer.
In this case, we can use the pigeon registers from this IP block in
order to do horizontal crop on the frame buffer processed by the
eLCDIF block.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>