MLK-22870 arm64: dts: add OCOTP clock
commit dd2aefcf94fa7c8aa24c79dd166db07374473783 from
https://source.codeaurora.org/external/imx/linux-imx

Use the kexec command to restart the root cell Linux, and the system
will hang up in imx8mm_soc_uid(). The reason is that ocotp root clk is
not enabled when the unique ID is read. Therefore, add CLK_OCOTP_ROOT to
init-on-array in root.dts to ensure that ocotp root clk in the root cell
Linux is always enabled.

Signed-off-by: Alice Guo <alice.guo@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>