MLK-24906: [8QM_MEK/8QXP_MEK]mxc:vpu_malone: report V4L2_COLORSPACE_DEFAULT before seq hdr found
commit  from
https://source.codeaurora.org/external/imx/linux-imx

the new gst need driver report V4L2_COLORSPACE_DEFAULT
before seq hdr is found,
otherwise the gpaly may fail to playback some stream.
To be compatible with the new version of gst,
change the behavior of try_fmt.

and the gst is adapted for kernel 5.10,
so this patch will only be pushed to 5.10,
won't be included in 5.4

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>