spi: spi-fsl-dspi: Use DMA for s32 controller in slave mode
commit c303d621b36de1a46943738c5c9207a1c04a1b41 from
https://source.codeaurora.org/external/autobsps32/linux

Switch to DMA for slave mode otherwise the controller is to slow to
feed TX FIFO and UNDERFLOW occurs frequently.
DMA can work only with 8 and 16 bits per word. 32bits per word
is not supported, this is a hardware limitation, so we
keep the master in TCFQ mode.
The slave supports a maximum of 1024 words per transfer
(this is the maximum of the internal buffer used by the
driver for dma). For more than 1024 words per transfer,
the driver splits the transfer in more dma transfers.
Between dma transfers, Underflow occurs in the slave and the
last transferred word is duplicated.

Issue: ALB-1129
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>