spi: spi-fsl-dspi: Use DMA for s32 controller in slave mode
commit  7b2378dc4b7b5e37d26b8867be71447acdec6d96 from
https://source.codeaurora.org/external/autobsps32/linux

Switch to DMA for slave mode otherwise the controller is to slow to
feed TX FIFO and UNDERFLOW occurs frequently.
DMA can work only with 8 and 16 bits per word. 32bits per word
is not supported, this is a hardware limitation, so we
keep the the master in TCFQ mode.
The slave supports a maximum of 1024 words per transfer
(this is the maximum of the internal buffer used by the
driver for dma). For more than 1024 words per transfer,
the driver splits the transfer in more dma transfers.
Between dma transfers, Underflow occurs in the slave and the
last transferred word is duplicated.

Issue: ALB-1129
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>