arm64: dts: ti: k3-j7200-main: Add hwspinlock node
commit 1d7a01c40840d844fb5d353c151f0ee0a7680c2f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The Main NavSS block on J7200 SoCs contains a HwSpinlock IP instance that
is same as the IP on AM65x and J721E SoCs. Add the DT node for this on
J7200 SoCs. The node is present within the Main NavSS block, and is added
as a child node under the main_navss interconnect node.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Praneeth Bajjuri <praneeth@ti.com>
Link: https://lore.kernel.org/r/20201026232637.15681-2-s-anna@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>