arm64: dts: ls1088ardb: add necessary DTS nodes for DPMAC2
commit f081dba5c13bb210106f7b63c748718fc587a57e from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

Annotate the external MDIO2 node and describe the 10GBASER PHY found on
the LS1088ARDB board and add a phy-handle for DPMAC2 to link it.
Also, add the internal PCS MDIO node for the internal MDIO buses found
on the LS1088A SoC along with its internal PCS PHY and link the
corresponding DPMAC to the PCS through the pcs-handle.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>