Documentation: fsl-dspi: Add slave pinctrl option
commit  3bc8047518f7f9885c161e7dbe1a1d7e32f91448 from
https://source.codeaurora.org/external/autobsps32/linux

The spi controller node in dts cand define both master
and slave pinctrl. The selection between them will be done based
on pinctrl-names. The default pinctrl will be loaded first
and will be used by the master. If the controller is configured
as slave (spi-slave property is added in the dts node), the
driver will look for the "slave" pinctrl and apply it if
found.

Issue: ALB-1129
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>