Added documentation for LLCE node in device tree.
commit  26b6895a3257d48de657e6cdbb83923a74765d45 from
https://source.codeaurora.org/external/autobsps32/linux

The documentation file explains the LLCE node, the interrupts associated
with the LLCE which are present in the device tree. It also shows the
region of memory that is used for LLCE registers and also for the shared
memory between LLCE and the CPUs.

Signed-off-by: Dorin Ionita <dorin.ionita@nxp.com>
Issue: ALB-4265
Signed-off-by: Meng Li <Meng.Li@windriver.com>