clk: imx: blk_ctrl: Use the imx_ccm_lock to protect on runtime PM callbacks
commit 70c20f7041a703eee3c494af08581c3520293656 from
https://source.codeaurora.org/external/imx/linux-imx

There is a chance the clocks might be gated while we're saving
the appropriate regs for runtime supend. There is no issue here by
using the imx_ccm_lock to also protect the reset bits. The alternative
would be to add a separate lock for the registers that hold reset bits.
That would complicate the logic since we would have to use different
locks for different registers.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>