soc: ti: Kconfig: Drop ARM64 SoC specific configs
commit e83b2358ab7ef0d39563b4e66233b356f99b7e77 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

With the integration of chip-id detection scheme in kernel[1], there
is no specific need to maintain multitudes of SoC specific config
options, discussed as per [2], we have deprecated the usage in other
places for v5.10-rc1. Drop the configuration for the follow on kernel.

[1] drivers/soc/ti/k3-socinfo.c commit 907a2b7e2fc7 ("soc: ti: add k3 platforms chipid module driver")

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>