MLK-24912-1 crypto: caam/jr - update jr_list during suspend/resume
commit c300818079e7363e148921e2b77a47ab1b725827 from
https://source.codeaurora.org/external/imx/linux-imx

The global driver_data.jr_list contains the list of active job rings
at a given moment.

Picking a JR is done using caam_jr_alloc(), which goes through this list
and chooses the JR with the least number of users ("tfm_count").

During the JR suspend/resume, this list must be updated to reflect that
the JR is no longer available - otherwise caam_jr_alloc() could return
a JR that has been suspended.

While this is rather a theoretical issue (i.e. was not met in practice),
it is a prerequisite for fixing the RNG failure met during suspend/resume.

Signed-off-by: Horia Geantă <horia.geanta@nxp.com>
Tested-by: Iuliana Prodan <iuliana.prodan@nxp.com>