drivers:fsl_qbman: remove device dependancy for link updates
commit 8754d4f2f5ad91eda9b4549ce83a0771c889a9ae from
https://source.codeaurora.org/external/imx/linux-imx

Qbman usdpaa driver is utilising the device's phy priv or net device priv
to store the eventfd context which is required to send signal to the
user space driver for link updates.
Storing the eventfd context in phy's priv is not a good idea.
In shared MAC case, kernel ethernet driver can fill phy's priv
and in that case fsl_qbman can over-write this priv, which may
cause driver mis-behaviour.

This patch introduces a new list of net devices which is
internal to fsl_qbman usdpaa driver to store the eventfd
contexts, thus removing the dependancy of device's priv.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
DPDK-2800