drm/vc4: hdmi: Don't poll for the infoframes status on setup
commit  321616069a68bd764024f801d73b25d6750ad5f3 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The infoframes are sent at a regular interval as a data island packet,
so we don't need to wait for them to be sent when we're setting them up.

However, we do need to poll when we're enabling since the we can't
update the packet RAM until it has been sent.

Let's add a boolean flag to tell whether we want to poll or not to
support both cases.

Suggested-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Reviewed-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20201203074624.721559-1-maxime@cerno.tech
Signed-off-by: Meng Li <Meng.Li@windriver.com>