soc: ti: k3-ringacc: Use correct device for allocation in RING mode
commit c7470fbb6ef3f6c64d9c0f654a9d0caaafb20f40 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 8c42379e40e2db4199ceeb6a6ef9fff73ff132cf upstream.

In RING mode the ringacc does not access the ring memory. In this access
mode the ringacc coherency does not have meaning.

If the ring is configured in RING mode, then the ringacc itself will not
access to the ring memory. Only the requester (user) of the ring is going
to read/write to the memory.
Extend the ring configuration parameters with a device pointer to be used
for DMA API when the ring is configured in RING mode.

Extending the ring configuration struct will allow per ring selection of
device to be used for allocation, thus allowing per ring coherency.

To avoid regression, fall back to use the ringacc dev in case the alloc_dev
is not provided.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Reviewed-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>