dmaengine: of-dma: Add support for optional router configuration callback
commit 94dd8301da49a1e505fedf37a729aed7fad01d96 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Additional configuration for the DMA event router might be needed for a
channel which can not be done during device_alloc_chan_resources callback
since the router information is not yet present for the drivers.

If there is a need for additional configuration for the channel if DMA
router is in use, then the driver can implement the device_router_config
callback.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>