dmaengine: ti: k3-udma: Initial support for K3 PKTDMA
commit 9bf45ca6eb7ef22fd6451cb16de2b5a2615ec82c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

One of the DMAs introduced with AM64 is the Packet DMA (PKTDMA).
It serves similar purpose as K3 UDMAP channels in packet mode, but with
notable differences, like tflow support and channels being allocated to
service specific peripherals.
The rings for the PKTDMA is integrated within the DMA itself instead of
using rings from the general purpose ringacc.

PKTDMA can be used to service PSI-L peripherals, similarly to
K3 UDMA channels.

Most of the driver code can be reused for PKTDMA tchan/rchan support but
new setup and allocation functions are needed to handle the differences
between the DMAs.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>