dmaengine: ti: k3-udma-glue: Add support for K3 PKTDMA
commit e230a5233902ab4014e60ea119fa6ca4b8a82e2a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This commit adds support for PKTDMA in k3-udma glue driver. Use new
psil_endpoint_config struct to get static data for a given channel or a
flow during setup.  Make sure that the RX flows being mapped to a RX
channel is within the range of flows that is been allocated to that RX
channel.

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>