drm/vc4: kms: Remove unassigned_channels from the HVS state
commit  7ccf97eb2d8439b8fbc72b1a1ebd7f376422be32 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The HVS state now has both unassigned_channels that reflects the
channels that are not used in the associated state, and the in_use
boolean for each channel that says whether or not a particular channel
is in use.

Both express pretty much the same thing, and we need the in_use variable
to properly track the commits, so let's get rid of unassigned_channels.

Suggested-by: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Reviewed-by: Thomas Zimmermann <tzimmermann@suse.de>
Link: https://patchwork.freedesktop.org/patch/msgid/20201204151138.1739736-6-maxime@cerno.tech
Signed-off-by: Meng Li <Meng.Li@windriver.com>