dmaengine: fsl-dpaa2-qdma: Fix the size of dma pools
commit 4b95fddb3e784019fd85836cc3bfe14653ed94a3 from
https://source.codeaurora.org/external/imx/linux-imx

In case of long format of qDMA command descriptor, there are one frame
descriptor, three entries in the frame list and two data entries. So the
size of dma_pool_create for these three fields should be the same with
the total size of entries respectively, or the contents may be overwritten
by the next allocated descriptor.

Signed-off-by: Guanhua Gao <guanhua.gao@nxp.com>