dmaengine: of-dma: Add support for optional router configuration callback
commit a5c660808a2fc027f88086a1d5e5a6d763638d36 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 4f910c035f38053ac8eb63a672c78862c535cd0f upstream.

Additional configuration for the DMA event router might be needed for a
channel which can not be done during device_alloc_chan_resources callback
since the router information is not yet present for the drivers.

If there is a need for additional configuration for the channel if DMA
router is in use, then the driver can implement the device_router_config
callback.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Link: https://lore.kernel.org/r/20201208090440.31792-8-peter.ujfalusi@ti.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
[s-anna@ti.com: cherry-pick commit '4f910c035f38' from v5.11]
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>