arm64: dts: ti: k3-am64-main: Add hwspinlock node
commit 4813c08e8641e12b01424930f4eb51dfd2412756 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The AM64x SoCs contains a HwSpinlock IP instance that is based on the
IP on AM65x SoCs. Add the DT node for this on AM64x SoCs. The node is
present within the the Main domain, and is added as a child node under
the cbass_main node.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>