ARM: dts: bcm2711: Add reserved memory template to hold firmware configuration
commit  c5b52730355a96713d292b7e72e8d8e818fbb09a from
https://github.com/raspberrypi/linux.git rpi-5.10.y

RPi4's co-processor will copy the board's bootloader[1] configuration
into memory for the OS to consume. Specifically, for the bootloader
configuration and upgrade user-space routines to query it through
nvmem's sysfs interface.

Introduce a reserved-memory area template for the co-processor to edit
before booting the system so as for Linux not to overwrite that memory
and to expose it as an nvmem device.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>

[1] https://www.raspberrypi.org/documentation/hardware/raspberrypi/bcm2711_bootloader_config.md
Signed-off-by: Meng Li <Meng.Li@windriver.com>