s32: spi: Add S32 node to the spi devicetree Documentation
commit 89d50a3f18fc06c4d4e518fb8b0e54faa269e300 from
https://source.codeaurora.org/external/autobsps32/linux

Including documenation for SPI slave support.
spi-slave property should be used for slave mode only.

The spi controller node in dts cand define both master
and slave pinctrl. The selection between them will be done based
on pinctrl-names. The default pinctrl will be loaded first
and will be used by the master. If the controller is configured
as slave (spi-slave property is added in the dts node), the
driver will look for the "slave" pinctrl and apply it if
found.

Issue: ALB-1129
Issue: ALB-5969
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>