spi: spi-zynqmp-gqspi: Resolved slab-out-of-bounds bug
commit  65a656500b44dc78a067bef32e1c11fd57621f19 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

During a transfer the driver filled the fifo with 4bytes,
even if the data that needs to be transfer is less that 4bytes.
This resulted in slab-out-of-bounds bug in KernelAddressSanitizer.

This patch resolves slab-out-of-bounds bug by filling the fifo
with the number of bytes that needs to transferred.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>