drm/vc4: hdmi: Don't register the CEC adapter if there's no interrupts
commit  1018347b9598d757d6087d4899dc63659f2e48de from
https://github.com/raspberrypi/linux.git rpi-5.10.y

We introduced the BCM2711 support to the vc4 HDMI controller with 5.10,
but this was lacking any of the interrupts of the CEC controller so we
have to deal with the backward compatibility.

Do so by simply ignoring the CEC setup if the DT doesn't have the
interrupts property.

Reviewed-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>