arm64: dts: ti: k3-j7200-som-p0: Add mailboxes to R5Fs
commit 379f93563e9494a094d84d3824950a623878c0f2 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[ Upstream commit 7a3b0c2ad3b0cfac5cb8820b8961f0956ff863fb ]

Add the required 'mboxes' property to all the R5F processors for the
TI J7200 common processor board. The mailboxes and some shared memory
are required for running the Remote Processor Messaging (RPMsg) stack
between the host processor and each of the R5Fs. The nodes are therefore
added in the common k3-j7200-som-p0.dtsi file so that all of these can
be co-located.

The chosen sub-mailboxes match the values used in the current firmware
images. This can be changed, if needed, as per the system integration
needs after making appropriate changes on the firmware side as well.

Note that any R5F Core1 resources are needed and used only when that
R5F cluster is configured for Split-mode.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>
Link: https://lore.kernel.org/r/20210111184554.6748-3-s-anna@ti.com
[s-anna@ti.com: cherry-pick commit '7a3b0c2ad3b0' from v5.12]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>