spi: cadence-quadspi: Wait at least 500 ms for direct reads
commit 25bc6a88878cc6389a6558af8b3d4156555c749e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 0920a32cf6f20467aa133a47b776ee782daa889f upstream.

When performing a direct read via DMA the timeout for completion is set
equal to the read length. This is fine for larger reads. For a small
read like the Read Status Register command, the timeout would be 1 or 2
milliseconds. This is not enough to cover the overhead needed in setting
up DMA.

Make sure the timeout is at least 500 ms to allow DMA ample time to
finish. For reads larger than 500 bytes, the timeout will continue to be
equal to the read length.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>