Revert "net: axienet: convert to phylink API"
commit  6ef08ce0e873e28c080ca93504d4b9a193cadb47 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This reverts commit f5203a3d9b25fa28a40bfc27c05a4020c6430f06.
There are reported link issues with SGMII and 1000Base-x MLO_AN_PHY
mode. To fix it we are in discussion with mainline and Russell
suggested extension to phylink framework and modification to axieth
driver which is work in progress. For now revert phylink support
till all xilinx usecases are supported and functional in the
mainline tree.

Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Link: https://lkml.org/lkml/2020/1/27/603
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>