tty: serial: linflexuart: Workaround for 'reboot' command freezing
commit 19aea4ac0800ed89ce601331c83fd201ef92114a from
https://source.codeaurora.org/external/autobsps32/linux

When executing the 'reboot' command, the linflex driver freezes
inside a DMA transfer function because of the DTFTFF bit of the
UARTSR register not being cleared. When operating using DMA, thus
in 'FIFO Mode', this bit is set and cleared by hardware to indicate
whether or not the TX FIFO is full.

This workaround writes a character to the device TX FIFO,
thus causing it send said character and in turn clear the
DTFTFF bit. The workaround is applied only after the boot sequence
has been completed, with the assumption that a standard boot process
does not take longer than 10 seconds to complete.

Issue: ALB-1009
Signed-off-by: Nica Dan <dan.nica@nxp.com>

tty: serial: linflexuart: Refactor set_termios

Issue: ALB-1848
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>