tty: serial: linflexuart: Fix fsl_linflexuart LINFBRR calculation
commit 42ff29103b86e3602bd2d02e80fefe012d3bbc17 from
https://source.codeaurora.org/external/autobsps32/linux

Baudrate is computed as LIN_CLK / (OSR × LDIV) when ROSE is 1 or LIN_CLK
/ (16 × LDIV) when ROSE is 0. LDIV is an unsigned fixed point number.
The mantissa is coded into 20 bits of LINIBRR and the fraction is coded on
4 bits of LINFBRR so setting the baudrate means setting LINFBRR and
LINIBRR.
LINIBRR is the integer part of LIN_CLK / f(ROSE) * baudrate and LINFBRR
is the fractional part of it multiplied by 16, where f(ROSE) can be 16
or OSR value.
LINFBRR wasn't correctly computed. If we consider divisr = LIN_CLK and
dividr = f(ROSE) * baudrate, LINFBRR should be the fractional part of
(divisr/dividr) multiplied by 16 and not (divisr%dividr)*16.

Issue: ALB-594
Signed-off-by: Larisa Grigore <Larisa.Grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>