LF-1571: mxc-jpeg: Fix decoding invalid buffer
commit 98d1e0561db18e105059adbf0a4f6c19bb452f19 from
https://source.codeaurora.org/external/imx/linux-imx

End of buffer was not properly detected in case of an invalid
buffer stream, such as a stream of null bytes, for example.
This was leading to a kernel oups.
Let get_byte return type be int, so it can return an error for
reaching end of stream.

While at it, replace 0 with NULL, for null pointer initialization.

Fixes: 36649dec0eb7 ("mxc-jpeg: Fix result independent of operands")
Signed-off-by: Mirela Rabulea <mirela.rabulea@oss.nxp.com>