spi: spi-cadence-quadspi: Add support for odd byte count read/write in DDR mode
commit  d0cdb6af5853b381ad4d4508a4209954a94aeb52 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Due to the nature of DDR protocol, an even number of bytes is always
transferred and also the least significant bit of the byte address
shall always be zero when using DDR protocol.
This patch adds support for read and write operations for the
requests with odd byte count or odd address or both. These kind of requests
are expected with UBIFS file system and mtd_stresstest.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>