arm64: dts: ti: k3-am642-sk: Add DDR carveout memory nodes for R5Fs
commit d67c7d533d584ed6c90b775deed1fb6d819ca0ab from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Two carveout reserved memory nodes each have been added for each of the
R5F remote processor devices within the MAIN domain on the TI AM642 SK
board. These nodes are assigned to the respective rproc device nodes
as well. The first region will be used as the DMA pool for the rproc
devices, and the second region will furnish the static carveout regions
for the firmware memory.

The current carveout addresses and sizes are defined statically for each
device. The R5F processors do not have an MMU, and as such require the
exact memory used by the firmwares to be set-aside. The firmware images
do not require any RSC_CARVEOUT entries in their resource tables to
allocate the memory for firmware memory segments. The addresses chosen
are the same as the respective processors on the AM64x EVM board to
maintain firmware compatibility between the two boards.

Note that the R5F1 carveouts are needed only if the R5F cluster is running
in Split (non Single-CPU) mode. The reserved memory nodes can be disabled
later on if there is no use-case defined to use the corresponding remote
processor.

Tested-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>