usb: dwc3: of-simple: Enable CCI support in dwc3 driver
commit  5453e8c07f8eebdea4fde1df9098e81c24d5eb1e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch adds CCI support in DWC3 driver when CCI is enabled in the
design. There are couple of registers which need to be modified for
making CCI enabled transfers in USB. This patch add the support for the
same.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>