net: dpaa2-mac: add backplane link mode support
commit 0823ea6685850a577b92dfb80b0a2575c9f290a7 from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

Add support for backplane link mode, which is, according to discussions
with NXP earlier in the year, is a mode where the OS (Linux) is able to
manage the PCS and Serdes itself.

This commit prepares the ground work for allowing 1G fiber connections
to be used with DPAA2 on the SolidRun CEX7 platforms.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>